+++ CMakeLists.txt
@@ -8,7 +8,7 @@ set(CMAKE_MODULE_PATH ${CMAKE_MODULE_PATH} ${PROJECT_SOURCE_DIR}/builds/cmake)
 
 # src
 set (cv_src 
-           src/cv.cpp)
+           src/CV.cpp)
 
 if(UNIX AND NOT APPLE)
     set(LINUX TRUE)
@@ -25,5 +25,5 @@ IF(MINGW)
 ENDIF(MINGW)
 
 
-add_executable(cv src/interpreter.cpp)
+add_executable(cv src/Interpreter.cpp)
 target_link_libraries(cv cv_bin)
