--- CV.hpp
@@ -1,19 +1,0 @@
-#ifndef CV_HPP
-    #define CV_HPP
-
-    #include <algorithm>
-    #include <functional>
-    #include <iostream>
-    #include <string>
-    #include <vector>
-    #include <unordered_map>
-    #include <cmath>
-    #include <inttypes.h>
-
-    #define __CV_NUMBER_NATIVE_TYPE double
-
-    namespace CV {
-
-    }
-
-#endif
