+++ src/cv.cpp
@@ -1,2420 +1,2438 @@
-#include "cv.hpp"
+#include <algorithm>
+#include <iostream>
+#include <algorithm>
+#include <cmath>
 
-
+#include "CV.hpp"
 namespace CV {
-	namespace Tools {
-		unsigned genItemCountId(){
-			static std::mutex genMutex;
-			static unsigned itemCountId = 1;
-			genMutex.lock();
-			auto id = ++itemCountId;
-			genMutex.unlock();
-			return id;
-		}
+    namespace Tools {
 
-		bool isNumber(const std::string &s){
-			return (s.find_first_not_of( "-.0123456789") == std::string::npos);
-		}
+        static void printList(const std::vector<std::string> &list){ // for debugging
+            std::cout << "START" << std::endl;
+            for(int i = 0 ; i < list.size(); ++i){
+                std::cout << "'" << list[i] << "'" << std::endl;
+            }
+            std::cout << "END\n" << std::endl;
+        }
 
-		double positive(double n){
-			return n < 0 ? n * -1.0 : n;
-		}
+        static std::string removeExtraSpace(const std::string &in){
+            std::string cpy;
+            for(int i = 0; i < in.length(); ++i){
+                if(i > 0 && in[i] == ' ' && in[i - 1] == ' '){
+                    continue;
+                }
+                if((i == 0 || i == in.length() - 1) && in[i] == ' '){
+                    continue;
+                }
+                cpy += in[i];
+            }
+            return cpy;
+        }
 
-		std::string filterEscapeChars(const std::string &input, bool remove){
-			std::string out;
+        bool isLineComplete(const std::string &input){
+            auto clean = Tools::removeExtraSpace(input);
+            int open = 0;
+            bool comm = false;
+            bool str = false;
+            for(int i = 0; i < input.length(); ++i){
+                char c = input[i];
+                // ++c;
+                if(c == '[' && !str) ++open;
+                if(c == ']' && !str) --open;
+                if(c == '#' && !str){
+                    comm = !comm;
+                    continue;
+                }
+                if(comm){
+                    continue;
+                }     
+                if(c == '\''){
+                    str = !str;
+                }   				
+            }	
+            return open == 0;		
+        }
 
-			for(int i = 0; i < input.size(); ++i){
-				char c = input[i];
-				if(c == '\n') { if(!remove){out += "\\n";} continue;} 
-				if(c == '\t') { if(!remove){out += "\\t";} continue;} 
-				out += input[i];
-			}
+        static bool isNumber(const std::string &s){
+            return (s.find_first_not_of( "-.0123456789") == std::string::npos);
+        }
 
-			return out;
-		}
+        static __CV_NUMBER_NATIVE_TYPE positive(__CV_NUMBER_NATIVE_TYPE n){
+            return n < 0 ? n * -1.0 : n;
+        }
 
-		bool isValidVarName(const std::string &s){
-			return (s.substr(0, 1).find_first_not_of( "-.0123456789" ) != std::string::npos);
-		}
+        static bool isString(const std::string &s){
+            return s.size() > 1 && s[0] == '\'' && s[s.length()-1] == '\'';
+        }  
 
-		bool isString(const std::string &s){
-			return s.size() > 1 && s[0] == '\'' && s[s.length()-1] == '\'';
-		}
+        static bool isList(const std::string &s){
+            if(s.length() < 3){
+                return false;
+            }
+            bool str = false;
+            for(unsigned i = 0; i < s.length(); ++i){
+                char c = s[i];
+                if(c == '\''){
+                    str = !str;
+                }     
+                if(!str && s[i] == ' '){
+                    return true;
+                }
+            }
+            return false;
+        }    
 
-		std::vector<std::string> split(const std::string &str, const char sep){
-			std::vector<std::string> tokens;
-			std::size_t start = 0, end = 0;
-			while ((end = str.find(sep, start)) != std::string::npos) {
-				tokens.push_back(str.substr(start, end - start));
-				start = end + 1;
-			}
-			tokens.push_back(str.substr(start));
-			return tokens;
-		}
+        static std::string removeTrailingZeros(double d){
+            size_t len = std::snprintf(0, 0, "%.8f", d);
+            std::string s(len+1, 0);
+            std::snprintf(&s[0], len+1, "%.8f", d);
+            s.pop_back();
+            s.erase(s.find_last_not_of('0') + 1, std::string::npos);
+            if(s.back() == '.') {
+                s.pop_back();
+            }
+            return s;
+        }
 
-		bool isList(const std::string &s){
+        static std::string compileList(const std::vector<std::string> &strings){
+            std::string out;
 
-			if(s.length() < 3){
-				return false;
-			}
+            for(int i = 0; i < strings.size(); ++i){
+                out += strings[i];
+                if(i < strings.size()-1){
+                    out += " ";
+                }
+            }
 
-			bool str = false;
+            return out;
+        }  
+    }
+}
 
-			for(unsigned i = 0; i < s.length(); ++i){
-				char c = s[i];
-				if(c == '\''){
-					str = !str;
-				}     
-				if(!str && s[i] == ' '){
-					return true;
-				}
-			}
+/*
 
-			return false;
+    TRAIT
 
-			// if()
-
-			// return s.size() > 1 && s[0] == '[' && s[s.length()-1] == ']' && s.find(' ') == -1;
-		}
-
-		std::string removeExtraSpace(const std::string &in){
-			std::string cpy;
-			for(int i = 0; i < in.length(); ++i){
-				if(i > 0 && in[i] == ' ' && in[i - 1] == ' '){
-					continue;
-				}
-				if((i == 0 || i == in.length() - 1) && in[i] == ' '){
-					continue;
-				}
-				cpy += in[i];
-			}
-			return cpy;
-		}
-
-
-		bool isLineComplete(const std::string &input){
-			auto clean = Tools::removeExtraSpace(input);
-			std::vector<std::string> tokens;
-			int open = 0;
-			bool comm = false;
-			bool str = false;
-			for(int i = 0; i < input.length(); ++i){
-				char c = input[i];
-				// ++c;
-				if(c == '[' && !str) ++open;
-				if(c == ']' && !str) --open;
-				if(c == '#' && !str){
-					comm = !comm;
-					continue;
-				}
-				if(comm){
-					continue;
-				}     
-				if(c == '\''){
-					str = !str;
-				}   				
-			}	
-			return open == 0;		
-		}
-
-		std::vector<std::string> parse(const std::string &input, std::string &error){
-			auto clean = Tools::removeExtraSpace(input);
-			std::vector<std::string> tokens;
-			int open = 0;
-			bool comm = false;
-			bool str = false;
-			bool mod = false;
-			std::string hold;
-			for(int i = 0; i < input.length(); ++i){
-				char c = input[i];
-				// ++c;
-				if(c == '[' && !str) ++open;
-				if(c == ']' && !str) --open;
-				if(c == '#' && !str){
-					comm = !comm;
-					continue;
-				}
-				if(comm){
-					continue;
-				}        
-				// escape characters
-				if(str && c == '\\' && i < input.length()-1){
-					char nc = input[i + 1];
-					// new line
-					if(nc == 'n'){
-						i += 1;
-						hold += "\n";
-						continue;
-					}
-					// tab
-					if(nc == 't'){
-						i += 1;
-						hold += "\t";
-						continue;
-					}	
-					// slash
-					if(nc == '\\'){
-						i += 1;
-						hold += "\\";
-						continue;
-					}
-					// single quote
-					if(nc == '\''){
-						i += 1;
-						hold += "'";
-						continue;
-					}														
-				}
-
-				if(c == '\''){
-					str = !str;
-				}   
-
-				hold += c;     
-				
-				mod = !str && open == 0 && c == ']' && i < input.length()-1 && (input[i+1] == '~' || input[i+1] == ':');
-				
-				if(!str && open == 0){
-					if((c == ']' && !mod) || c == ' ' || i ==  input.length()-1){
-						if(hold != " "){
-							tokens.push_back(Tools::removeExtraSpace(hold));
-						}
-						mod = false;
-						hold = "";
-					}
-				}
-			}
-			if(open != 0){
-				error = "Bad parenthesis";
-				return {};
-			}
-			return tokens;
-		}
-
-		int getModifiers(const std::string &input, std::unordered_map<std::string, std::string> &mods){
-
-			int i = input.size()-1;
-			int lastFound = i;
-			std::string hold;
-			while(i > 0){
-				char c = input[i];
-				
-				if(c == ']'){
-					break;
-				}else
-				if(c == ':' || c == '~'){
-					mods[std::string()+c] = std::string(hold.rbegin(), hold.rend());
-					hold = "";
-					--i;
-					lastFound = i;
-					continue;
-				}
-				
-				hold += c;
-				--i;
-			}
-
-			return lastFound;
-
-		}
-
-		std::vector<std::string> getCleanTokens(std::string &input, std::string &error){
-			auto tokens = Tools::parse(input, error);
-
-			if(error.size() > 0){
-				return {};
-			}
-
-			if(tokens.size() == 1 && tokens[0][0] == '[' && tokens[0][tokens[0].length()-1] == ']'){
-				auto trimmed = tokens[0].substr(1, tokens[0].length() - 2);
-				return getCleanTokens(trimmed, error);
-			}	
-
-			return tokens;
-		}
-	}
+*/
+CV::Trait::Trait(){
+    this->type = CV::TraitType::SPECIFIC;
+    this->action = [](Item *subject, const std::string &value, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+        return std::shared_ptr<CV::Item>(NULL);
+    };
 }
 
 CV::Item::Item(){
-	type = ItemTypes::NIL;
-	data = NULL;
-	size = 0;
-	id = Tools::genItemCountId();	
-	name = ""; // anonymous	
-	temporary = false;
+    registerTraits();
+    this->type = CV::ItemTypes::NIL;
 }
 
-void CV::Item::clear(){	
-	if(type != ItemTypes::NIL){
-		free(data);
-	}
-	type = ItemTypes::NIL;
-	data = NULL;
-	size = 0;
+void CV::Item::registerTrait(uint8_t type, const std::function<std::shared_ptr<CV::Item>(Item *subject, const std::string &value, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx)> &action){
+    Trait trait;
+    trait.action = action;
+    trait.type = type;
+    this->traitsAny[type] = trait;
+}    
+
+void CV::Item::registerTrait(const std::string &name, const std::function<std::shared_ptr<CV::Item>(Item *subject, const std::string &value, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx)> &action){
+    Trait trait;
+    trait.name = name;
+    trait.action = action;
+    this->traits[name] = trait;
 }
 
-void CV::Item::write(void *data, size_t size, int type){
-	if(this->data != NULL){
-		clear();
-	}
-	this->data = malloc(size);
-	this->size = size;
-	this->type = type;
-	memcpy(this->data, data, size);
+std::shared_ptr<CV::Item> CV::Item::runTrait(const std::string &name, const std::string &value, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+    auto &trait = this->traits.find(name)->second;
+    return trait.action( this, value, cursor, ctx);
 }
 
-std::shared_ptr<CV::Item> CV::Item::copy(){
-	auto r = std::make_shared<Item>(Item());
-	r->type = type;
-	r->size = size;
-	r->data = malloc(size);
-	memcpy(r->data, data, size);
-	return r;
+std::shared_ptr<CV::Item> CV::Item::runTrait(uint8_t type, const std::string &value, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+    return this->traitsAny.find(type)->second.action(this, value, cursor, ctx);
+}    
+
+bool CV::Item::hasTrait(const std::string &name){
+    auto it = this->traits.find(name);
+    if(it == this->traits.end()){
+        return false;
+    }
+    return true;
 }
 
-void CV::Item::registerProperty(const std::string &name, const std::shared_ptr<Item> &v, bool onRealContext){
-	if(onRealContext){
-		auto ctx = this->findRealContext();
-		if(ctx != NULL){
-			ctx->members[name] = v;
-		}
-	}else{
-		this->members[name] = v;
-	}
+bool CV::Item::hasTrait(uint8_t type){
+    auto it = this->traitsAny.find(type);
+    if(it == this->traitsAny.end()){
+        return false;
+    }
+    return true;
 }
 
-void CV::Item::deregisterProperty(const std::string &name){
-	this->members.erase(name);
-}	
-
-std::shared_ptr<CV::Item> CV::Item::findProperty(const std::string &name){
-	auto it = this->members.find(name);
-	if(it == this->members.end()){
-		return this->head.get() == NULL ? std::make_shared<CV::Item>(CV::Item()) : head->findProperty(name);
-	}
-	return it->second;		
-}	
-
-CV::Item *CV::Item::findRealContext(){
-	if(this->temporary){
-		return this->head.get() == NULL ? NULL : this->head->findRealContext();
-	}
-	return this;
+bool CV::Item::isEq(std::shared_ptr<CV::Item> &item){
+    return (this->type == CV::ItemTypes::NIL && item->type == CV::ItemTypes::NIL) || this == item.get();
 }
 
-CV::Item *CV::Item::findContext(const std::string &name){
-	auto it = this->members.find(name);
-	if(it == this->members.end()){
-		return this->head.get() == NULL ? NULL : this->head->findContext(name);
-	}
-	return this;
+std::shared_ptr<CV::Item> CV::Item::copy(bool deep) const {
+    return std::make_shared<CV::Item>(CV::Item());
 }
 
-std::shared_ptr<CV::Item> CV::Item::findAndSetProperty(const std::string &name, const std::shared_ptr<CV::Item> &v){
-	auto ctx = this->findContext(name);
-	if(ctx == NULL) return std::make_shared<CV::Item>(CV::Item());
-	ctx->members[name] = v;
-	v->name = name;
-	return ctx->members[name];	
-}
+ bool CV::Item::clear(bool deep){
+    if(this->type == CV::ItemTypes::NIL){
+        return false;
+    }
 
-std::shared_ptr<CV::Item> CV::Item::getProperty(const std::string &name){
-	auto it = this->members.find(name);
-	if(it == this->members.end()){
-		return std::make_shared<Item>(Item());
-	}
-	return it->second;
-}
-
-std::string CV::Item::str(bool singleLine) const {
-	switch(type){
-		case ItemTypes::NUMBER: {
-			std::ostringstream oss;
-			oss << std::setprecision(8) << std::noshowpoint << *static_cast<double*>(data);
-			return singleLine ? CV::Tools::filterEscapeChars(oss.str()) : oss.str();
-		} break;
-		case ItemTypes::NIL: {
-			return "nil";
-		} break;
-		case ItemTypes::CONTEXT: {
-			return "CONTEXT";
-		} break;			
-		default:
-			//TODO: HANDLE ERROR
-			std::exit(1);
-	}
-}
-
-CV::Item::operator std::string() const {
-	return str();
-}	
-
-CV::NumberType::NumberType(){
-	this->type = ItemTypes::NUMBER;
-	this->data = new double(0);
-}
-
-CV::NumberType::~NumberType(){
-	this->clear();
-}
-
-void CV::NumberType::set(int n){
-	double _n = (double)n;
-	memcpy(this->data, &_n, sizeof(_n));
-}
-
-void CV::NumberType::set(float n){
-	double _n = (double)n;
-	memcpy(this->data, &_n, sizeof(_n));
-}
-
-void CV::NumberType::set(double n){
-	double _n = (double)n;
-	memcpy(this->data, &_n, sizeof(_n));
-}
-
-double CV::NumberType::get(){
-	double n;
-	memcpy(&n, this->data, sizeof(n));
-	return n;
-}
-
-std::shared_ptr<CV::Item> CV::NumberType::copy(){
-	auto copy = std::make_shared<NumberType>(NumberType());
-	copy->set(get());
-	return std::static_pointer_cast<Item>(copy);
-}	
-
-std::string CV::NumberType::str(bool singleLine) const {
-	switch(type){
-		case ItemTypes::NUMBER: {
-			std::ostringstream oss;
-			oss << std::setprecision(8) << std::noshowpoint << *static_cast<double*>(data);
-			return oss.str();
-		} break;
-		case ItemTypes::NIL: {
-			return "nil";
-		} break;
-		default: 
-			// TODO: HANDLE ERROR
-			std::exit(1);
-	}		
+    return true;
 }
 
 
-CV::StringType::StringType(){
-	this->type = ItemTypes::STRING;
-	this->literal = "";
-	this->registerProperty("length", create(0));
-}
-std::shared_ptr<CV::Item> CV::StringType::copy(){
-	auto copy = std::shared_ptr<CV::StringType>(new CV::StringType());
-	copy->literal = this->literal;
-	return std::static_pointer_cast<Item>(copy);
-}
+/*
 
-std::shared_ptr<CV::Item> CV::StringType::get(int index){
-	if(index < 0 || index >= this->literal.size()){
-		return std::make_shared<CV::Item>(CV::Item());
-	}
-	auto result = std::make_shared<CV::StringType>(CV::StringType());	
-	result->set(std::string()+this->literal[index]);
-	return result;
+    INTERRUPT
+
+*/
+CV::Interrupt::Interrupt(uint8_t intType){
+    this->type = CV::ItemTypes::INTERRUPT;
+    this->intType = intType;
 }
 
-void CV::StringType::set(const std::string &v){
-	this->literal = v;
-	std::static_pointer_cast<NumberType>(this->getProperty("length"))->set((int)this->literal.size());
-}	
+std::shared_ptr<CV::Item> CV::Interrupt::copy(bool deep) const {
+    auto item = std::make_shared<CV::Interrupt>(this->intType);
+    item->payload = this->payload;
+    return std::static_pointer_cast<CV::Item>(item);
+}  
 
-std::string CV::StringType::str(bool singleLine) const {
-	switch(type){
-		case ItemTypes::STRING: {
-			return singleLine ? CV::Tools::filterEscapeChars(this->literal) : this->literal;
-		} break;
-		case ItemTypes::NIL: {
-			return "nil";
-		} break;
-		default: 
-			// TODO: HANDLE ERROR
-			std::exit(1);
-	}		
+bool CV::Interrupt::hasPayload(){
+    return this->payload.get() != NULL;
+}
+
+void CV::Interrupt::setPayload(std::shared_ptr<CV::Item> &item){
+    this->payload = item;
+}
+
+std::shared_ptr<CV::Item> CV::Interrupt::getPayload(){
+    return this->payload;
 }
 
 
-CV::ListType::ListType(int n){
-	type = ItemTypes::LIST;
-	for(int i = 0; i < n; ++i){
-		this->list.push_back(create(0));
-	}
-	this->registerProperty("length", create((int)this->list.size()));
-}
-void CV::ListType::build(const std::vector<std::shared_ptr<CV::Item>> &objects){
-	this->list = objects;
-	std::static_pointer_cast<CV::NumberType>(this->getProperty("length"))->set((int)this->list.size());
-}
-void CV::ListType::add(const std::shared_ptr<CV::Item> &item){
-	this->list.push_back(item);
-	std::static_pointer_cast<CV::NumberType>(this->getProperty("length"))->set((int)this->list.size());
-}
-std::shared_ptr<CV::Item> CV::ListType::get(int index){
-	if(index < 0 || index >= this->list.size()){
-		return std::make_shared<Item>(Item());
-	}
-	return this->list[index];
+/*
+
+    NUMBER
+
+*/
+CV::Number::Number(){
+    registerTraits();
 }
 
-std::shared_ptr<CV::Item> CV::ListType::copy(){
-	return std::make_shared<CV::Item>(CV::Item());
-}
-std::string CV::ListType::str(bool singleLine) const {
-	switch(type){
-		case ItemTypes::LIST: {
-			std::string _params;
-			std::string buff = "";
-			for(int i = 0; i < this->list.size(); ++i){
-				buff += list[i]->str(singleLine);
-				if(i < this->list.size()-1){
-					buff += " ";
-				}                
-			}
-			return singleLine ? CV::Tools::filterEscapeChars("["+buff+"]") : "["+buff+"]";
-		} break;
-		case ItemTypes::NIL: {
-			return "nil";
-		} break;
-		default: 
-			std::exit(1);
-	}
+CV::Number::Number(__CV_NUMBER_NATIVE_TYPE v){
+    registerTraits();        
+    set(v);
 }
 
+bool CV::Number::isEq(std::shared_ptr<CV::Item> &item){
+    return item->type == this->type && this->n == std::static_pointer_cast<CV::Number>(item)->n;
+}    
 
-
-CV::FunctionType::FunctionType(){
-	this->type = ItemTypes::FUNCTION;
-	this->body = "";
-	this->inner = false;
-	this->lambda = [](const std::vector<std::shared_ptr<Item>> &operands, std::shared_ptr<Item> &ctx, Cursor *cursor){
-		return std::make_shared<Item>(Item());
-	};  
-}
-CV::FunctionType::FunctionType(const std::function<std::shared_ptr<Item>(const std::vector<std::shared_ptr<Item>> &operands, std::shared_ptr<Item> &ctx, Cursor *cursor)> &lambda,
-			const std::vector<std::string> &params){
-				this->type = ItemTypes::FUNCTION;
-				this->lambda = lambda;
-				this->params = params;
-				this->inner = true;
-			}
-
-void CV::FunctionType::set(const std::string &body, const std::vector<std::string> &params){
-	this->lambda = [body](const std::vector<std::shared_ptr<Item>> &operands, std::shared_ptr<Item> &ctx, Cursor *cursor){
-		return eval(body, ctx, cursor);
-	};    
-	this->body = body;
-	this->inner = false;
-	this->params = params;
-}
-
-std::shared_ptr<CV::Item> CV::FunctionType::copy(){
-	return std::make_shared<CV::Item>(CV::Item());
-}
-
-std::string CV::FunctionType::str(bool singleLine) const {
-	switch(type){
-		case ItemTypes::FUNCTION: {
-			if(this->inner){
-				return "[BINARY]";
-			}
-			std::string _params;
-			for(int i = 0; i < params.size(); ++i){
-				_params += params[i];
-				if(i < params.size()-1){
-					_params += " ";
-				}
-			}
-			return singleLine ? CV::Tools::filterEscapeChars("fn ["+_params+"] -> ["+body+"]") : "fn ["+_params+"] -> ["+body+"]";
-		} break;
-		case ItemTypes::NIL: {
-			return "nil";
-		} break;
-		default:
-			std::exit(1);
-	}
+void CV::Number::registerTraits(){
+    this->registerTrait("is-neg", [](Item *subject, const std::string &value, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+        if(subject->type == CV::ItemTypes::NIL){
+            return std::make_shared<Item>();
+        }
+        return std::static_pointer_cast<CV::Item>(std::make_shared<CV::Number>(CV::Number(static_cast<Number*>(subject)->n < 0)));
+    });
+    this->registerTrait("inv", [](Item *subject, const std::string &value, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+        if(subject->type == CV::ItemTypes::NIL){
+            return std::make_shared<Item>();
+        }
+        return std::static_pointer_cast<CV::Item>(std::make_shared<CV::Number>(CV::Number(static_cast<Number*>(subject)->n * -1)));
+    });    
+    this->registerTrait("is-odd", [](Item *subject, const std::string &value, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+        if(subject->type == CV::ItemTypes::NIL){
+            return std::make_shared<Item>();
+        }
+        return std::static_pointer_cast<CV::Item>(std::make_shared<CV::Number>(CV::Number( (int)static_cast<Number*>(subject)->n % 2 != 0)));
+    });   
+    this->registerTrait("is-even", [](Item *subject, const std::string &value, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+        if(subject->type == CV::ItemTypes::NIL){
+            return std::make_shared<Item>();
+        }
+        return std::static_pointer_cast<CV::Item>(std::make_shared<CV::Number>(CV::Number( (int)static_cast<Number*>(subject)->n % 2 == 0)));
+    });                       
 }
 
+std::shared_ptr<CV::Item> CV::Number::copy(bool deep) const {
+    return std::make_shared<CV::Number>(CV::Number(this->n));
+}    
 
-
-
-std::shared_ptr<CV::Item> CV::InterruptType::copy(){
-	auto r = std::make_shared<InterruptType>();
-	r->intype = intype;
-	return std::static_pointer_cast<Item>(r);		
+void CV::Number::set(const __CV_NUMBER_NATIVE_TYPE v){
+    this->type = CV::ItemTypes::NUMBER;
+    this->n = v;
 }
 
-CV::InterruptType::InterruptType(int intype, std::shared_ptr<CV::Item> payload){
-	this->type = ItemTypes::INTERRUPT;
-	this->intype = intype;
-	this->payload = payload;
+__CV_NUMBER_NATIVE_TYPE CV::Number::get(){
+    return this->n;
 }
 
-std::string CV::InterruptType::str(bool singleLine) const {
-	switch(type) {
-		case ItemTypes::INTERRUPT: {
-			switch(intype){
-				case InterruptTypes::STOP: {
-					return "[INTERRUPT-STOP]";
-				} break;
-				case InterruptTypes::SKIP: {
-					return "[INTERRUPT-SKIP]";
-				} break;
-				case InterruptTypes::RET: {
-					return "[INTERRUPT-RET]";
-				} break;
-				default: {
-					return "[INTERRUPT]";
-				} break;                             
-			}
-		} break;
-		case ItemTypes::NIL: {
-			return "nil";
-		} break;
-		default: 
-			std::exit(1);
-	}
+/*
+
+    LIST
+
+*/
+CV::List::List(){
+    registerTraits();
 }
 
-
-
-CV::ProtoType::ProtoType(){
-	this->type = ItemTypes::PROTO;
-	this->registerProperty("length", create(0));
+CV::List::List(const std::vector<std::shared_ptr<CV::Item>> &list, bool toCopy){
+    registerTraits();        
+    set(list, toCopy);
 }
 
-void CV::ProtoType::populate(const std::vector<std::string> &names, const std::vector<std::shared_ptr<Item>> &values, std::shared_ptr<Item> &ctx, Cursor *cursor){
-	for(unsigned i = 0; i < names.size(); ++i){
-		auto name = names[i];
-		this->registerProperty(names[i], values[i]);
-	}
-	std::static_pointer_cast<NumberType>(this->getProperty("length"))->set((int)members.size()-1);
+void CV::List::registerTraits(){
+    this->registerTrait(CV::TraitType::ANY_NUMBER, [](Item *subject, const std::string &value, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+        if(subject->type == CV::ItemTypes::NIL){
+            return std::make_shared<Item>();
+        }
+        auto list = static_cast<List*>(subject); 
+        auto index = std::stod(value);
+        if(index < 0 || index >= list->data.size()){
+            cursor->setError("'List|*ANY_NUMBER*': Index("+Tools::removeTrailingZeros(index)+") is out of range(0-"+std::to_string(list->data.size())+")");
+            return std::make_shared<CV::Item>(CV::Item());
+        }
+        return list->get(index); 
+    });        
+    this->registerTrait("n", [](Item *subject, const std::string &value, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+        if(subject->type == CV::ItemTypes::NIL){
+            return std::make_shared<Item>();
+        }
+        auto list = static_cast<List*>(subject); 
+        return std::static_pointer_cast<CV::Item>(std::make_shared<CV::Number>(CV::Number(list->data.size())));
+    });
+    this->registerTrait("reverse", [](Item *subject, const std::string &value, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+        if(subject->type == CV::ItemTypes::NIL){
+            return std::make_shared<Item>();
+        }
+        std::vector<std::shared_ptr<CV::Item>> result;
+        auto list = static_cast<List*>(subject); 
+        for(int i = 0; i < list->data.size(); ++i){
+            result.push_back(list->data[list->data.size()-1 - i]);
+        }
+        return std::static_pointer_cast<CV::Item>(std::make_shared<CV::List>(CV::List(result)));
+    });                           
 }
 
-
-void CV::ProtoType::add(const std::string &name, std::shared_ptr<Item> &item, std::shared_ptr<Item> &ctx, Cursor *cursor){ 
-	this->registerProperty(name, item);
-	std::static_pointer_cast<NumberType>(this->getProperty("length"))->set((int)members.size()-1);
+std::shared_ptr<CV::Item> CV::List::copy(bool deep) const {
+    auto nlist = std::make_shared<CV::List>(CV::List());
+    for(int i = 0; i < this->data.size(); ++i){
+        nlist->data.push_back(deep ? this->data[i]->copy() : this->data[i]);
+    }        
+    return std::static_pointer_cast<CV::Item>(nlist);
 }
 
-std::vector<std::string> CV::ProtoType::getKeys(){
-	std::vector<std::string> names;
-	for(auto &it : this->members){
-		if(it.first == "length"){
-			continue;
-		}
-		names.push_back(it.first);
-	}
-	return names;
+void CV::List::set(const std::vector<std::shared_ptr<CV::Item>> &list, bool toCopy){
+    this->type = CV::ItemTypes::LIST;
+    for(int i = 0; i < list.size(); ++i){
+        this->data.push_back(toCopy ? list[i]->copy() : list[i]);
+    }        
 }
 
-
-std::string CV::ProtoType::str(bool singleLine) const {
-	auto addTab = [](unsigned n){
-		std::string tabs;
-		for(unsigned i = 0; i < n; ++i){
-			tabs += "\t";
-		}
-		return tabs;
-	};
-	unsigned maxWidth = 0;
-	std::vector<std::string> names;
-	std::vector<std::string> values;
-	std::vector<int> ids;
-	for(auto &it : this->members){
-		maxWidth = std::max((unsigned)it.first.length(), maxWidth);
-		names.push_back(it.first);
-		values.push_back(it.second->str(true));
-		ids.push_back(it.second->id);
-	}		
-	std::string out = singleLine ? "PROTO " : "PROTO\n";
-	for(unsigned i = 0; i < names.size(); ++i){
-		if(singleLine){
-			out += names[i]+std::string(std::string()+": ")+values[i];
-			if(i < names.size()-1){
-				out += ", ";
-			}
-		}else{
-			out += names[i];
-			out += addTab(maxWidth);
-			out += std::string(": ") + CV::Tools::filterEscapeChars(values[i])+"\n";
-		}
-	}
-	return out;
+std::shared_ptr<CV::Item> CV::List::get(size_t index) const {
+    return this->data[index];
 }
 
+bool CV::List::clear(bool deep){
+    if(this->type == CV::ItemTypes::NIL){
+        return false;
+    }
+    return true;
+}   
 
-std::shared_ptr<CV::Item> CV::create(double n){
-	auto result = std::make_shared<CV::Item>(CV::Item());
-	double v = n;
-	result->write(&v, sizeof(v), ItemTypes::NUMBER);
-	return result;	
+
+/*
+
+    STRING
+
+*/
+CV::String::String(){
+    // Empty constructor makes this string NIL
+    registerTraits();
 }
 
-std::shared_ptr<CV::Item> CV::createContext(const std::shared_ptr<CV::Item> &head, bool temporary){
-	auto result = std::make_shared<CV::Item>(CV::Item());
-	result->type = CV::ItemTypes::CONTEXT;
-	result->head = head;
-	result->temporary = temporary;
-	return result;	
+CV::String::String(const std::string &str){
+    registerTraits();
+    this->set(str);
+}    
+
+void CV::String::registerTraits(){ 
+    this->registerTrait(CV::TraitType::ANY_NUMBER, [](Item *subject, const std::string &value, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+        if(subject->type == CV::ItemTypes::NIL){
+            return std::make_shared<Item>();
+        }
+        auto str = static_cast<String*>(subject);
+        auto index = std::stod(value);
+        if(index < 0 || index >= str->data.size()){
+            cursor->setError("'String|*ANY_NUMBER*': Index("+Tools::removeTrailingZeros(index)+") is out of range(0-"+std::to_string(str->data.size())+")");
+            return std::make_shared<CV::Item>(CV::Item());
+        }
+        return std::static_pointer_cast<CV::Item>(std::make_shared<CV::String>(CV::String(  std::string("")+str->data[index]  )));
+    });           
+    this->registerTrait("n", [](Item *subject, const std::string &value, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+        if(subject->type == CV::ItemTypes::NIL){
+            return std::make_shared<Item>();
+        }
+        auto str = static_cast<String*>(subject);
+        return std::static_pointer_cast<CV::Item>(std::make_shared<CV::Number>(CV::Number( str->data.size() )));
+    });
+    this->registerTrait("list", [](Item *subject, const std::string &value, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+        if(subject->type == CV::ItemTypes::NIL){
+            return std::make_shared<Item>();
+        }
+        std::vector<std::shared_ptr<CV::Item>> result;
+        auto str = static_cast<String*>(subject);
+
+        for(int i = 0; i < str->data.size(); ++i){
+            result.push_back(std::static_pointer_cast<CV::Item>(std::make_shared<CV::String>(CV::String(  std::string("")+str->data[i]  ))));
+        }
+
+
+        return std::static_pointer_cast<CV::Item>(std::make_shared<CV::List>(CV::List(result)));
+    });        
+    this->registerTrait("reverse", [](Item *subject, const std::string &value, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+        if(subject->type == CV::ItemTypes::NIL){
+            return std::make_shared<Item>();
+        }
+        std::string result;
+        auto str = static_cast<String*>(subject);
+
+        for(int i = 0; i < str->data.size(); ++i){
+            result += str->data[str->data.size()-1 - i];
+        }
+        return std::static_pointer_cast<CV::Item>(std::make_shared<CV::String>(CV::String(  result  )));
+
+    });                           
 }
 
-std::shared_ptr<CV::ListType> CV::createList(int n){
-	return std::make_shared<CV::ListType>(CV::ListType(n));	
+bool CV::String::isEq(std::shared_ptr<CV::Item> &item){
+    if(this->type != item->type || this->data.size() != std::static_pointer_cast<CV::String>(item)->data.size()){
+        return false;
+    }
+    return this->data == std::static_pointer_cast<CV::String>(item)->data;
+} 
+
+std::shared_ptr<CV::Item> CV::String::copy(bool deep) const {
+    auto copied = std::make_shared<CV::String>(CV::String());
+    copied->set(this->data);
+    return copied;
 }
 
-std::shared_ptr<CV::ProtoType> CV::createProto(){
-	return std::make_shared<CV::ProtoType>(CV::ProtoType());	
-}
+void CV::String::set(const std::string &str){
+    this->type = CV::ItemTypes::STRING;
+    this->data = str;
+} 
 
-std::shared_ptr<CV::StringType> CV::createString(const std::string &str){
-	auto result = std::make_shared<CV::StringType>(CV::StringType());	
-	result->set(str);
-	return result;
+std::string &CV::String::get(){     
+    return this->data;   
 }
 
 
+/*
 
+    FunctionConstraints
 
-std::shared_ptr<CV::Item> CV::infer(const std::string &input, std::shared_ptr<CV::Item> &ctx, Cursor *cursor){
-	if(cursor->error){
-		return std::make_shared<Item>(Item());
-	}	
-	if(Tools::isNumber(input) && input.find(' ') == -1){
-		auto result = std::make_shared<Item>(Item());
-		auto v = std::stod(input);
-		result->write(&v, sizeof(v), ItemTypes::NUMBER);
-		return result;
-	}else
-	// String
-	if(Tools::isString(input)){
-		auto result = std::make_shared<StringType>(StringType());
-		result->literal = input.substr(1, input.length() - 2);
-		return result;        
-	}else
-	if(Tools::isList(input)){
-		auto result = createList(0);
-		std::string _input = input;
-		if(_input.length() > 1 && _input[0] == '[' && _input[_input.length()-1] == ']'){
-			_input = _input.substr(1, _input.length() - 2);		
-		}		
-		std::string error;
-		auto tokens = Tools::parse(_input, error);
-		if(error.length() > 0){
-			cursor->setError(error);
-			return std::make_shared<Item>(Item());
-		}
-		for(int i = 0; i < tokens.size(); ++i){			
-			auto item = eval(tokens[i], ctx, cursor);
-			if(cursor->error){
-				return std::make_shared<Item>(Item());
-			}				
-			result->add(item);
-		}
-		return result;
-	}else{
-		return std::make_shared<Item>(Item());
-	}
+*/
+CV::FunctionConstraints::FunctionConstraints(){
+    enabled = true;
+    allowNil = true;
+    allowMisMatchingTypes = true;
+    useMinParams = false;
+    useMaxParams = false;
 }
 
-
-
-static std::shared_ptr<CV::Item> evalParams(std::vector<std::string> &tokens, int from, std::shared_ptr<CV::Item> &ctx, CV::Cursor *cursor){
-	std::string input = "";
-	for(int i = from; i < tokens.size(); ++i){
-		input += tokens[i];
-		if(i < tokens.size()-1){
-			input += " ";
-		}
-	}
-	return CV::eval(input, ctx, cursor);
+void CV::FunctionConstraints::setExpectType(uint8_t type){
+    this->expectedTypes.push_back(type);
 }
 
-static std::vector<std::string> compileParams(const std::vector<std::string> &tokens, int from){
-	std::vector<std::string> compiled;
-	for(int i = from; i < tokens.size(); ++i){
-		compiled.push_back(tokens[i]);
-	}
-	return compiled;
+void CV::FunctionConstraints::setExpectedTypeAt(int pos, uint8_t type){
+    this->expectedTypeAt[pos] = type;
 }
 
-static std::string compileParams(const std::vector<std::string> &tokens){
-	std::string compiled;
-	for(int i = 0; i < tokens.size(); ++i){
-		compiled += tokens[i];
-		if(i < tokens.size()-1) compiled += " ";
-	}
-	return compiled;
+void CV::FunctionConstraints::clearExpectedTypes(){
+    this->expectedTypeAt.clear();
+    this->expectedTypes.clear();
 }
 
-static std::shared_ptr<CV::Item> runFunction(std::shared_ptr<CV::Item> &fn, const std::vector<std::string> &tokens, std::shared_ptr<CV::Item> &ctx, CV::Cursor *cursor){
-	
-	auto func = static_cast<CV::FunctionType*>(fn.get());
-	bool useParams = func->params.size() > 0;
-
-	if(useParams && func->params.size() != tokens.size()){
-		// std::cout << "fn '" << "': expected [" << func->params.size() << "] operands, received [" << tokens.size() << "]" << std::endl;
-		return std::make_shared<CV::Item>(CV::Item());
-	}
-
-	std::vector<std::shared_ptr<CV::Item>> operands;
-
-	bool useFuncParams = func->params.size() >= tokens.size();
-	for(int i = 0; i < tokens.size(); ++i){
-
-		auto ev = eval(tokens[i], ctx, cursor);
-
-		if(cursor->error){
-			return std::make_shared<CV::Item>(CV::Item());
-		}	
-
-		if(ev->type == CV::ItemTypes::INTERRUPT){
-			auto interrupt = std::static_pointer_cast<CV::InterruptType>(ev);
-			if(interrupt->intype == CV::InterruptTypes::RET){
-				ev = interrupt->payload;
-			}else{
-				ev = std::make_shared<CV::Item>(CV::Item());
-			}
-			return ev;
-		}
-
-		if(useParams){
-			// ev->name = func->params[i-1];
-		}
-		ctx->registerProperty(useFuncParams ?  func->params[i] : "__param"+std::to_string(i), ev);
-		operands.push_back(ev);
-	} 
-	auto result = func->lambda(operands, ctx, cursor); 
-	if(cursor->error){
-		return std::make_shared<CV::Item>(CV::Item());
-	}	
-	return result;	
+void CV::FunctionConstraints::setMaxParams(unsigned n){
+    if(n == 0){
+        useMaxParams = false;
+        return;
+    }
+    enabled = true;
+    maxParams = n;
+    useMaxParams = true;
 }
 
-
-std::shared_ptr<CV::Item> CV::eval(const std::string &input, std::shared_ptr<CV::Item> &ctx, CV::Cursor *cursor){
-
-	if(cursor->error){
-		return std::make_shared<Item>(Item());
-	}	
-
-	if(input.length() == 0){
-		cursor->setError("There's nothing here...");
-		return std::make_shared<CV::Item>(Item()); 
-	}
-	std::string error;
-	auto tokens = Tools::parse(input, error);
-	if(error.size() > 0){
-		cursor->setError(error);
-		return std::make_shared<CV::Item>(Item());
-	}
-
-	auto imp = tokens[0];
-
-	// find modifiers if any
-	std::string scopeMod = "";
-	std::string namer = "";
-	if(imp.rfind(":") != -1 || imp.rfind("~") != -1){
-		std::unordered_map<std::string, std::string> mods;
-		int cutOff = Tools::getModifiers(imp, mods);
-
-		imp = imp.substr(0, cutOff+1);
-		tokens[0] = imp;
-		if(mods.count(":") > 0){
-			scopeMod = mods[":"];
-		}
-		if(mods.count("~") > 0){
-			namer = mods["~"];	
-		}		
-	}
-
-	// chained
-	if(tokens.size() > 1 && tokens[0].length() > 1 && tokens[0][0] == '[' && tokens[0][tokens[0].length()-1] == ']'){
-		auto last = CV::createList(0);
-		// chained are evaluated and results compiled into a single list
-		for(int i = 0; i < tokens.size(); ++i){
-			auto imp = tokens[i];
-			auto obj = eval(imp, ctx, cursor);
-			if(namer.size() > 0){
-				ctx->registerProperty(namer, obj, true);
-				obj->name = namer;
-			}				
-			if(obj->type == ItemTypes::INTERRUPT){
-				auto interrupt = std::static_pointer_cast<InterruptType>(obj);
-				if(interrupt->intype == InterruptTypes::RET && interrupt->payload->type != ItemTypes::NIL){
-					return interrupt->payload;
-				}else{
-					return interrupt;
-				}
-			}			
-			last->add(obj);
-			if(cursor->error){
-				return std::make_shared<Item>(Item());
-			}				
-		}
-
-		return last;
-	}else
-	// single
-	if(tokens.size() == 1 && tokens[0].length() > 1 && tokens[0][0] == '[' && tokens[0][tokens[0].length()-1] == ']'){
-		imp = tokens[0].substr(1, tokens[0].length() - 2);		
-		auto obj = eval(imp, ctx, cursor);
-
-		if(namer.size() > 0){
-			ctx->registerProperty(namer, obj, true);
-			obj->name = namer;
-		}			
-		if(cursor->error){
-			return std::make_shared<Item>(Item());
-		}			
-		if(obj->type == ItemTypes::LIST && scopeMod.size() > 0 && Tools::isNumber(scopeMod)){
-			auto list = std::static_pointer_cast<ListType>(obj);
-			auto item = list->get(std::stod(scopeMod));
-			if(namer.size() > 0){
-				ctx->registerProperty(namer, item, true);
-				item->name = namer;
-			}
-			return item;
-		}else
-		if(obj->type == ItemTypes::STRING && scopeMod.size() > 0 && Tools::isNumber(scopeMod)){
-			auto str = std::static_pointer_cast<StringType>(obj);
-			auto item = str->get(std::stod(scopeMod));
-			if(namer.size() > 0){
-				ctx->registerProperty(namer, item, true);
-				item->name = namer;
-			}					
-			return item;
-		}else
-		if(scopeMod.size() > 0){
-			auto inobj = obj->getProperty(scopeMod);
-			if(inobj->type == ItemTypes::FUNCTION){
-				auto params = compileParams(tokens, 1);
-				auto tctx = createContext(ctx, true);
-				tctx->registerProperty("_", obj);
-				auto result = runFunction(obj, params, tctx, cursor);
-				if(cursor->error){
-					return std::make_shared<Item>(Item());
-				}			
-				if(namer.size() > 0){
-					tctx->registerProperty(namer, result, true);
-					inobj->name = namer;
-				}							
-				return result;
-			}else{
-				if(namer.size() > 0){
-					ctx->registerProperty(namer, inobj, true);
-					inobj->name = namer;
-				}				
-				return inobj;
-			}
-		}
-		return obj;
-	}
-
-	// Reserved imperative
-	if(isReserved(imp)){
-		// NIL
-		if(imp == "nil"){
-			return std::make_shared<Item>(Item());
-		}else
-		// SET declares a variable in the curret scope
-		if(imp == "set"){
-			auto name =  tokens[1];
-			auto val = eval(tokens[2], ctx, cursor);
-			if(cursor->error){
-				return std::make_shared<Item>(Item());
-			}			
-			if(val->type == ItemTypes::INTERRUPT){
-				auto interrupt = std::static_pointer_cast<InterruptType>(val);
-				if(interrupt->intype == InterruptTypes::RET){
-					val = interrupt->payload;
-				}else{
-					val = std::make_shared<Item>(Item());
-				}
-			}
-			val->name = name;
-			ctx->registerProperty(name, val);
-			return val;
-		}else
-		if(imp == "rset"){
-			// this is a dirty hack to allow modifying items within _.
-			// the solution is probably to redesign Item to be multi-type
-			// that would require fully dropping STL
-			std::string scopeMod;
-			std::string imp = tokens[1];
-			if(tokens[1].rfind(":") != -1 || tokens[1].rfind("~") != -1){
-				std::unordered_map<std::string, std::string> mods;
-				int cutOff = Tools::getModifiers(imp, mods);
-				imp = imp.substr(0, cutOff+1);
-				if(mods.count(":") > 0){
-					scopeMod = mods[":"];
-				}		
-			}	
-			auto ownerCtx = ctx;
-			auto who = eval(imp, ownerCtx, cursor);
-			if(cursor->error){
-				return std::make_shared<Item>(Item());
-			}			
-			if(scopeMod.size() > 0 && who->getProperty(scopeMod)->type != ItemTypes::NIL){
-				imp = scopeMod;
-				ownerCtx = who;
-			}
-
-			auto to = eval(tokens[2], ctx, cursor);
-			if(cursor->error){
-				return std::make_shared<Item>(Item());
-			}			
-
-			auto mutated = ownerCtx->findAndSetProperty(imp, to);
-	
-			return mutated;
-		}else
-		// PROTO builds a prototype item: the other first class citizen of canvas (other one being list)
-		if(imp == "proto"){
-			auto proto = std::make_shared<ProtoType>(ProtoType());
-
-			std::vector<std::string> names;
-			std::vector<std::shared_ptr<Item>> values;
-
-			for(unsigned i = 1; i < tokens.size(); ++i){
-
-				auto &token = tokens[i];
-
-				auto fc = token.find(":");
-
-				std::string varname = token.substr(0, fc);
-				std::string varvalue = token.substr(fc+1, token.length()-fc-1);
-
-
-				names.push_back(varname);
-				auto result = eval(varvalue, ctx, cursor);
-				result->registerProperty("_", proto);
-				if(cursor->error){
-					return std::make_shared<Item>(Item());
-				}					
-				values.push_back(result);				
-			}
-
-			proto->populate(names, values, ctx, cursor);	
-
-			return proto;		
-		}else
-		// FUNCTION
-		if(imp == "fn"){
-			auto largs = tokens[1];
-			if(largs.length() > 1 && largs[0] == '[' && largs[largs.length()-1] == ']'){
-				largs = largs.substr(1, largs.length() - 2);
-			}
-			std::string error;
-			auto args = Tools::parse(largs, error);
-			if(error.size() > 0){
-				cursor->setError(error);
-				return std::make_shared<Item>(Item());
-			}
-			auto fn = std::make_shared<FunctionType>(FunctionType());
-			fn->set(tokens[2], args);
-			return fn;
-		}else
-		// RET
-		if(imp == "ret"){
-			auto interrupt = std::make_shared<InterruptType>(InterruptType(InterruptTypes::RET));
-			// eval parameters
-			if(tokens.size() > 1){
-				interrupt->payload = evalParams(tokens, 1, ctx, cursor);
-				if(cursor->error){
-					return std::make_shared<Item>(Item());
-				}				
-			}
-			return std::static_pointer_cast<Item>(interrupt);  
-		}else
-		// SKIP (skips current execution ie inside a loop, next iteration [do true [[std:print test][skip]]]))
-		if(imp == "skip"){
-			auto interrupt = std::make_shared<InterruptType>(InterruptType(InterruptTypes::SKIP));
-			// eval parameters
-			if(tokens.size() > 1){
-				interrupt->payload = evalParams(tokens, 1, ctx, cursor);
-				if(cursor->error){
-					return std::make_shared<Item>(Item());
-				}					
-			}
-			return std::static_pointer_cast<Item>(interrupt);  
-		}else
-		// STOP (stops current execution [do [cond] [[std:print test][stop]]])
-		if(imp == "stop"){
-			auto interrupt = std::make_shared<InterruptType>(InterruptType(InterruptTypes::STOP));
-			// eval parameters
-			if(tokens.size() > 1){
-				interrupt->payload = evalParams(tokens, 1, ctx, cursor);	
-				if(cursor->error){
-					return std::make_shared<Item>(Item());
-				}				
-			}
-			return std::static_pointer_cast<Item>(interrupt);  
-		}else
-		// IF conditional statement ([if [cond] [branch-true][branch-false]])
-		if(imp == "if"){
-			if(tokens.size() < 3){
-				cursor->setError("operator 'if': expects at least 2 operands");
-				return std::make_shared<Item>(Item());
-			}
-			if(tokens.size() > 4){
-				cursor->setError("operator 'if': no more than 3 operands");
-				return std::make_shared<Item>(Item());
-			}
-			auto cond = eval(tokens[1], ctx, cursor);
-			if(cursor->error){
-				return std::make_shared<Item>(Item());
-			}			
-			bool t = cond->type == ItemTypes::NUMBER && *static_cast<double*>(cond->data) > 0;
-			if(t){
-				auto tctx = createContext(ctx, true);
-				auto trueBranch = eval(tokens[2], tctx, cursor);
-				if(cursor->error){
-					return std::make_shared<Item>(Item());
-				}
-				return trueBranch;
-			}else
-			if(tokens.size() > 3){
-				auto tctx = createContext(ctx, true);
-				auto falseBranch = eval(tokens[3], tctx, cursor);
-				if(cursor->error){
-					return std::make_shared<Item>(Item());
-				}				
-				return falseBranch;
-			}
-		}else
-		// for loop procedure ([for [start][cond][modifier] [code]...[code]])
-		if(imp == "for"){
-			if(tokens.size() < 5){
-				cursor->setError("operator 'for': expects at least 5 operands");
-				return std::make_shared<Item>(Item());
-			}	
-			bool t = false;
-			auto ctctx = createContext(ctx, true);
-			auto last = std::make_shared<Item>(Item());;
-
-			// start
-			auto part = eval(tokens[1], ctctx, cursor);
-			if(cursor->error){
-				return std::make_shared<Item>(Item());
-			}				
-
-			do {
-				auto cond = eval(tokens[2], ctctx, cursor);
-				if(cursor->error){
-					return last;
-				}		
-				t = cond->type == ItemTypes::NUMBER && *static_cast<double*>(cond->data) > 0;
-				if(t){
-					auto ntokens = compileParams(compileParams(tokens, 4));
-					auto run = eval(ntokens, ctctx, cursor);
-					last = run;
-					if(cursor->error){
-						return last;
-					}
-					if(run->type == ItemTypes::INTERRUPT){
-						return std::static_pointer_cast<InterruptType>(run)->payload;
-					}
-					auto modifier = eval(tokens[3], ctctx, cursor);
-					if(cursor->error){
-						return last;
-					}							
-				}
-			} while(t);
-			return last;
-		}else		
-		// DO loop procedure ([do [cond] [code]...[code]])
-		if(imp == "do"){
-			if(tokens.size() < 3){
-				cursor->setError("operator 'do': expects at least 2 operands");
-				return std::make_shared<Item>(Item());
-			}	
-			bool t = false;
-			auto ctctx = createContext(ctx, true);
-			auto last = std::make_shared<Item>(Item());;
-			do {
-				auto cond = eval(tokens[1], ctctx, cursor);
-				if(cursor->error){
-					return std::make_shared<Item>(Item());
-				}		
-				t = cond->type == ItemTypes::NUMBER && *static_cast<double*>(cond->data) > 0;
-				if(t){
-					auto tctx = createContext(ctctx, true);
-					auto ntokens = compileParams(compileParams(tokens, 2));
-					auto run = eval(ntokens, tctx, cursor);
-					last = run;
-					if(cursor->error){
-						return last;
-					}
-					if(run->type == ItemTypes::INTERRUPT){
-						return std::static_pointer_cast<InterruptType>(run)->payload;
-					}
-				}
-			} while(t);
-			return last;
-		}else
-		// ITER goes through a list ([iter [1 2 3 4] [code]...[code]]) ([iter [proto a:1 b:2 c:3] [code]...[code]])
-		if(imp == "iter"){
-			if(tokens.size() < 3){
-				cursor->setError("operator 'iter': expects at least 2 operands");
-				return std::make_shared<Item>(Item());
-			}
-			auto iteratable = eval(tokens[1], ctx, cursor);
-
-
-			if(cursor->error){
-				return std::make_shared<Item>(Item());
-			}
-
-			int readFrom = 2;
-
-			if(iteratable->type != ItemTypes::LIST && iteratable->type != ItemTypes::PROTO){
-				cursor->setError("operator 'iter': expects an iterable as first operands");
-				return std::make_shared<Item>(Item());
-			}
-	
-			switch(iteratable->type){
-				case ItemTypes::PROTO: {
-					auto proto = static_cast<ProtoType*>(iteratable.get());
-					std::shared_ptr<ListType> last = std::make_shared<ListType>(ListType());
-					auto names = proto->getKeys();
-					for(int i = 0; i < names.size(); ++i){
-						auto current = proto->members[names[i]];
-						auto fctx = createContext(ctx, false);
-						if(iteratable->name.size() > 0){
-							fctx->registerProperty(iteratable->name, current);
-						}
-						auto item = evalParams(tokens, readFrom, fctx, cursor);
-						if(cursor->error){
-							return std::make_shared<Item>(Item());
-						}
-						if(item->type == ItemTypes::INTERRUPT){
-							auto interrupt = std::static_pointer_cast<InterruptType>(item);
-							if(interrupt->intype == InterruptTypes::RET){
-								last->add(interrupt->payload);
-							}else
-							if(interrupt->intype == InterruptTypes::SKIP){
-								last->add(interrupt->payload);
-								continue;
-							}else
-							if(interrupt->intype == InterruptTypes::STOP){
-								last->add(interrupt->payload);
-								return last;
-							}
-						}else
-						if(item->type == ItemTypes::FUNCTION){
-							auto funct = std::static_pointer_cast<FunctionType>(item);
-							std::vector<std::shared_ptr<Item>> params = { current };
-							last->add(funct->lambda(params, fctx, cursor));
-						}else{
-							last->add(item);
-						}
-					}
-					return last;					
-				} break;
-				case ItemTypes::LIST: {
-					auto list = static_cast<ListType*>(iteratable.get());
-					std::shared_ptr<ListType> last = std::make_shared<ListType>(ListType());
-					for(int i = 0; i < list->list.size(); ++i){
-						auto current = list->list[i];
-						auto fctx = createContext(ctx, true);
-						if(iteratable->name.size() > 0){
-							fctx->registerProperty(iteratable->name, current);
-						}
-						auto item = evalParams(tokens, readFrom, fctx, cursor);
-						if(cursor->error){
-							return std::make_shared<Item>(Item());
-						}
-						if(item->type == ItemTypes::INTERRUPT){
-							auto interrupt = std::static_pointer_cast<InterruptType>(item);
-							if(interrupt->intype == InterruptTypes::RET){
-								last->add(interrupt->payload);
-							}else
-							if(interrupt->intype == InterruptTypes::SKIP){
-								last->add(interrupt->payload);
-								continue;
-							}else
-							if(interrupt->intype == InterruptTypes::STOP){
-								last->add(interrupt->payload);
-								return last;
-							}
-						}else
-						if(item->type == ItemTypes::FUNCTION){
-							auto funct = std::static_pointer_cast<FunctionType>(item);
-							std::vector<std::shared_ptr<Item>> params = { current };
-							last->add(funct->lambda(params, fctx, cursor));
-						}else{
-							last->add(item);
-						}
-					}
-					return last;
-				};				
-			}
-
-		}
-	}else{
-	// Defined symbol
-		auto var = ctx->findProperty(imp);	
-		if(var->type == ItemTypes::LIST && scopeMod.length() > 0 && Tools::isNumber(scopeMod)){
-			auto list = std::static_pointer_cast<ListType>(var);
-			auto item = list->get(std::stod(scopeMod));
-			if(namer.size() > 0){
-				ctx->registerProperty(namer, item, true);
-				item->name = namer;
-			}					
-			return item;
-		}else
-		if(var->type == ItemTypes::STRING && scopeMod.length() > 0 && Tools::isNumber(scopeMod)){
-			auto str = std::static_pointer_cast<StringType>(var);
-			auto item = str->get(std::stod(scopeMod));
-			if(namer.size() > 0){
-				ctx->registerProperty(namer, item, true);
-				item->name = namer;
-			}					
-			return item;
-		}else		
-		if(var->type == ItemTypes::NUMBER || var->type == ItemTypes::LIST || var->type == ItemTypes::PROTO || var->type == ItemTypes::STRING){
-			if(scopeMod.length() == 0 && tokens.size() > 1){ // perhaps this is part of a list
-				auto params = compileParams(tokens);
-				auto v = infer(params, ctx, cursor);
-				if(cursor->error){
-					return std::make_shared<Item>(Item());
-				}
-				if(namer.size() > 0){
-					ctx->registerProperty(namer, v, true);
-					v->name = namer;
-				}						
-				return v;
-			}
-			if(scopeMod.length() > 0){
-				auto obj = var->getProperty(scopeMod);
-				if(obj->type == ItemTypes::FUNCTION){
-					auto params = compileParams(tokens, 1);
-					auto tctx = createContext(ctx, true);
-					if(obj->getProperty("_")->type != ItemTypes::NIL){
-						tctx->registerProperty("_", var);
-					}
-					auto result = runFunction(obj, params, tctx, cursor);
-					if(cursor->error){
-						return std::make_shared<Item>(Item());
-					}
-					if(result->type == ItemTypes::INTERRUPT){
-						auto interrupt = std::static_pointer_cast<InterruptType>(result);
-						if(interrupt->intype == InterruptTypes::RET){
-							return interrupt->payload;
-						}
-					}else{						
-						if(scopeMod.size() > 0){
-							obj->name = scopeMod;
-						}							
-						if(namer.size() > 0){
-							tctx->registerProperty(namer, obj, true);
-							obj->name = namer;
-						}										
-						return result;
-					}
-				}else{		
-					if(scopeMod.size() > 0){
-						obj->name = scopeMod;
-					}							
-					if(namer.size() > 0){
-						ctx->registerProperty(namer, obj, true);
-						obj->name = namer;
-					}							
-					return obj;
-				}
-			}
-			if(namer.size() > 0){
-				ctx->registerProperty(namer, var, true);
-				var->name = namer;
-			}				
-			return var;
-
-		}else
-		if(var->type == ItemTypes::FUNCTION){
-			auto params = compileParams(tokens, 1);
-			auto tctx = createContext(ctx, true);
-			auto result = runFunction(var, params, tctx, cursor);
-			if(cursor->error){
-				return std::make_shared<Item>(Item());
-			}
-
-			if(namer.size() > 0){
-				tctx->registerProperty(namer, result, true);
-				result->name = namer;
-			}				
-			return result;			
-		}else{
-			auto params = compileParams(tokens);
-			auto v = infer(params, ctx, cursor);
-
-			if(cursor->error){
-				return std::make_shared<Item>(Item());
-			}
-
-			if(namer.size() > 0){
-				ctx->registerProperty(namer, v, true);
-				v->name = namer;
-			}	
-
-			if(v->type == ItemTypes::LIST && scopeMod.length() > 0 && Tools::isNumber(scopeMod)){
-				auto list = std::static_pointer_cast<ListType>(v);
-				auto item = list->get(std::stod(scopeMod));					
-				return item;
-			}else
-			if(v->type == ItemTypes::STRING && scopeMod.length() > 0 && Tools::isNumber(scopeMod)){
-				auto str = std::static_pointer_cast<StringType>(v);
-				auto item = str->get(std::stod(scopeMod));				
-				return item;
-			}else{
-				if(cursor->error){
-					return std::make_shared<Item>(Item());
-				}
-				return v;
-			}
-		}
-	}
-
-
-	return std::make_shared<Item>(Item());
-
-
+void CV::FunctionConstraints::setMinParams(unsigned n){
+    if(n == 0){
+        useMinParams = false;
+        return;
+    }
+    enabled = true;
+    minParams = n;
+    useMinParams = true;
 }
 
+/*
 
+    FUNCTION
 
-
-static bool doTypesMatch(const std::vector<std::shared_ptr<CV::Item>> &items){
-	int lastType = -1;
-	for(int i = 0; i < items.size(); ++i){
-		auto item = items[i];
-		if(lastType != -1 && item->type != lastType){
-			return false;
-		}
-		lastType = item->type;
-	}
-	return true;
+*/
+CV::Function::Function(){
+    registerTraits();
 }
 
-static bool doSizesMatch(const std::vector<std::shared_ptr<CV::Item>> &items){
-	int lastSize = -1;
-	for(int i = 0; i < items.size(); ++i){
-		auto item = std::static_pointer_cast<CV::ListType>(items[i]);
-		if(lastSize != -1 && item->list.size() != lastSize){
-			return false;
-		}
-		lastSize = item->list.size();
-	}
-	return true;
+CV::Function::Function(const std::vector<std::string> &params, const std::string &body, bool variadic){
+    registerTraits();
+    set(params, body, variadic);
 }
 
-static int isThereNil(const std::vector<std::shared_ptr<CV::Item>> &items){
-	for(int i = 0; i < items.size(); ++i){
-		auto item = items[i];
-		if(item->type == CV::ItemTypes::NIL){
-			return i;
-		}
-	}
-	return -1;
+CV::Function::Function(const std::vector<std::string> &params, const std::function<std::shared_ptr<CV::Item> (const std::vector<std::shared_ptr<CV::Item>> &params, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx)> &fn, bool variadic){
+    registerTraits();
+    set(params, fn, variadic);
+}    
+
+
+void CV::Function::registerTraits(){
+    this->registerTrait("is-variadic", [](Item *subject, const std::string &value, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+        if(subject->type == CV::ItemTypes::NIL){
+            return std::make_shared<Item>();
+        }
+        auto func = static_cast<Function*>(subject);
+        return std::static_pointer_cast<CV::Item>(std::make_shared<CV::Number>(CV::Number( func->variadic )));
+    });
+
+    this->registerTrait("is-binary", [](Item *subject, const std::string &value, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+        if(subject->type == CV::ItemTypes::NIL){
+            return std::make_shared<Item>();
+        }
+        auto func = static_cast<Function*>(subject);
+        return std::static_pointer_cast<CV::Item>(std::make_shared<CV::Number>(CV::Number( func->binary )));
+    }); 
+
+    this->registerTrait("n-args", [](Item *subject, const std::string &value, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+        if(subject->type == CV::ItemTypes::NIL){
+            return std::make_shared<Item>();
+        }
+        auto func = static_cast<Function*>(subject);
+        return std::static_pointer_cast<CV::Item>(std::make_shared<CV::Number>(CV::Number( func->params.size()  )));            
+    });   
+
+    this->registerTrait("proto", [](Item *subject, const std::string &value, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+        if(subject->type == CV::ItemTypes::NIL){
+            return std::make_shared<Item>();
+        }
+        auto func = static_cast<Function*>(subject);
+        return std::static_pointer_cast<CV::Item>(std::make_shared<CV::String>(CV::String(  func->binary ? "[]" : func->body  )));
+        
+    });                       
+                        
 }
 
-std::string CV::printContext(CV::Item *ctx, bool ignoreInners){
-	std::string output = "CONTEXT["+std::to_string(ctx->id)+"]\n";
-
-	int fcol = 5;
-	int scol = 5;
-
-	for(auto &it : ctx->members){
-		const auto item = it.second;
-		if(ignoreInners && item->type == CV::ItemTypes::FUNCTION && std::static_pointer_cast<CV::FunctionType>(item)->inner){
-			continue;
-		}
-		while(it.first.length() > fcol){
-			fcol += 5;
-		}
-	}
-
-	for(auto &it : ctx->members){
-		const auto item = it.second;
-		if(ignoreInners && item->type == CV::ItemTypes::FUNCTION && std::static_pointer_cast<CV::FunctionType>(item)->inner){
-			continue;
-		} 
-		output += it.first+"["+std::to_string(it.second->id)+"]";
-		int toAdd = fcol - it.first.length();
-		for(int i = 0; i < toAdd; ++i){
-			output += " ";
-		}
-		output += "->";
-		for(int i = 0; i < scol; ++i){
-			output += " ";
-		}    
-		output += it.second->type == CV::ItemTypes::STRING ?  it.second->str(true)+"\n" : "'"+it.second->str(true)+"'\n";
-	}
-
-	if(ctx->head.get()){
-		output += std::string()+"\n\n"+printContext(ctx->head.get(), ignoreInners);
-	}
-
-	return output;
+void CV::Function::set(const std::vector<std::string> &params, const std::string &body, bool variadic){
+    this->body = body;
+    this->params = params;
+    this->type = CV::ItemTypes::FUNCTION;
+    this->fn = [](const std::vector<std::shared_ptr<CV::Item>> &params, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+        return std::shared_ptr<CV::Item>(NULL);
+    };
+    this->binary = false;
+    this->variadic = variadic;
 }
 
-
-
-void CV::registerEmbeddedOperators(std::shared_ptr<CV::Item> &ctx){
-
-	/* 
-
-		OR
-
-	*/
-	ctx->registerProperty("or", std::make_shared<FunctionType>(FunctionType([](const std::vector<std::shared_ptr<Item>> &operands, std::shared_ptr<Item> &ctx, Cursor *cursor){
-			if(operands.size() < 2){
-				cursor->setError("operator 'or': expects at least 2 operand");
-				return std::make_shared<Item>(Item());						
-			}
-
-
-			for(int i = 0; i < operands.size(); ++i){			
-				if(operands[i]->type == ItemTypes::NUMBER && *static_cast<double*>(operands[i]->data) > 0){
-					return CV::create(1);		
-				}
-			}		
-			
-			return CV::create(0);
-			
-			return std::make_shared<Item>(Item());				
-		}, {}
-	))); 
-
-	/* 
-
-		NOR
-
-	*/
-	ctx->registerProperty("nor", std::make_shared<FunctionType>(FunctionType([](const std::vector<std::shared_ptr<Item>> &operands, std::shared_ptr<Item> &ctx, Cursor *cursor){
-			if(operands.size() < 2){
-				cursor->setError("operator 'nor': expects at least 2 operand");
-				return std::make_shared<Item>(Item());						
-			}
-
-
-			for(int i = 0; i < operands.size(); ++i){			
-				if(operands[i]->type == ItemTypes::NUMBER && *static_cast<double*>(operands[i]->data) > 0){
-					return CV::create(0);		
-				}
-			}		
-			
-			return CV::create(1);
-			
-			return std::make_shared<Item>(Item());				
-		}, {}
-	))); 
-
-
-
-	/* 
-
-		AND
-
-	*/
-	ctx->registerProperty("and", std::make_shared<FunctionType>(FunctionType([](const std::vector<std::shared_ptr<Item>> &operands, std::shared_ptr<Item> &ctx, Cursor *cursor){
-			if(operands.size() < 2){
-				cursor->setError("operator 'and': expects at least 2 operand");
-				return std::make_shared<Item>(Item());						
-			}
-
-			for(int i = 0; i < operands.size(); ++i){			
-				if(operands[i]->type != ItemTypes::NUMBER || *static_cast<double*>(operands[i]->data) <= 0){
-					return CV::create(0);		
-				}
-			}		
-			
-			return CV::create(1);
-			
-			return std::make_shared<Item>(Item());				
-		}, {}
-	))); 
-
-
-	/* 
-
-		NAND
-
-	*/
-	ctx->registerProperty("nand", std::make_shared<FunctionType>(FunctionType([](const std::vector<std::shared_ptr<Item>> &operands, std::shared_ptr<Item> &ctx, Cursor *cursor){
-			if(operands.size() < 2){
-				cursor->setError("operator 'nand': expects at least 2 operand");
-				return std::make_shared<Item>(Item());						
-			}
-
-			for(int i = 0; i < operands.size(); ++i){			
-				if(operands[i]->type != ItemTypes::NUMBER || *static_cast<double*>(operands[i]->data) <= 0){
-					return CV::create(1);		
-				}
-			}		
-			
-			return CV::create(0);
-			
-			return std::make_shared<Item>(Item());				
-		}, {}
-	))); 	
-
-	/* 
-
-		ADDITION
-
-	*/
-	ctx->registerProperty("+", std::make_shared<FunctionType>(FunctionType([](const std::vector<std::shared_ptr<Item>> &operands, std::shared_ptr<Item> &ctx, Cursor *cursor){
-			if(operands.size() < 2){
-				cursor->setError("operator '+': no operands provided");
-				return std::make_shared<Item>(Item());						
-			}
-			int nilInd = isThereNil(operands);
-			if(nilInd != -1){
-				cursor->setError("operator '+': argument("+std::to_string(nilInd+1)+") is nil");
-				return std::make_shared<Item>(Item());				
-			}
-			if(!doTypesMatch(operands)){
-				cursor->setError("operator '+': mismatching types");
-				return std::make_shared<Item>(Item());
-			}
-
-			if(operands[0]->type == ItemTypes::NUMBER){
-				auto result = std::make_shared<Item>(Item());
-				double v = *static_cast<double*>(operands[0]->data);
-				for(unsigned i = 1; i < operands.size(); ++i){
-					v += *static_cast<double*>(operands[i]->data);
-				}
-				result->write(&v, sizeof(v), ItemTypes::NUMBER);
-				return result;
-			}else
-			if(operands[0]->type == ItemTypes::LIST){
-				if(!doSizesMatch(operands)){
-					cursor->setError("operator '+': lists of mismatching sizes");
-					return std::make_shared<Item>(Item());						
-				}
-				auto firstItem = std::static_pointer_cast<ListType>(operands[0]);
-
-				if(firstItem->list.size() == 0){
-					cursor->setError("operator '+': empty lists");
-					return std::make_shared<Item>(Item());						
-				}
-
-				auto result = std::make_shared<ListType>(ListType( firstItem->list.size() ));
-				for(unsigned i = 0; i < operands.size(); ++i){
-					auto list = std::static_pointer_cast<ListType>(operands[i]);
-					for(unsigned j = 0; j < list->list.size(); ++j){
-						auto from = list->list[j];
-						auto to = result->list[j];
-						double v = *static_cast<double*>(from->data) + *static_cast<double*>(to->data);
-						to->write(&v, sizeof(v), ItemTypes::NUMBER);
-					}
-				}
-				return std::static_pointer_cast<Item>(result);
-			}else
-			if(operands[0]->type == ItemTypes::STRING){
-
-				std::string str = "";
-
-				for(unsigned i = 0; i < operands.size(); ++i){
-					auto current = std::static_pointer_cast<StringType>(operands[i]);
-					str += current->literal;
-				}
-
-				auto result = std::make_shared<StringType>(StringType());
-				result->set(str);
-
-
-				return std::static_pointer_cast<Item>(result);
-			}else{
-				cursor->setError("operator '+': invalid type");
-				return std::make_shared<Item>(Item());				
-			}		
-			
-			return std::make_shared<Item>(Item());				
-		}, {}
-	))); 
-
-	/*
-	
-		SUBSTRACTION
-
-	*/
-	ctx->registerProperty("-", std::make_shared<FunctionType>(FunctionType([](const std::vector<std::shared_ptr<Item>> &operands, std::shared_ptr<Item> &ctx, Cursor *cursor){
-			if(operands.size() < 2){
-				cursor->setError("operator '-': no operands provided");
-				return std::make_shared<Item>(Item());						
-			}
-			int nilInd = isThereNil(operands);
-			if(nilInd != -1){
-				cursor->setError("operator '-': argument("+std::to_string(nilInd+1)+") is nil");
-				return std::make_shared<Item>(Item());				
-			}
-			if(!doTypesMatch(operands)){
-				cursor->setError("operator '-': mismatching types");
-				return std::make_shared<Item>(Item());
-			}
-
-			if(operands[0]->type == ItemTypes::NUMBER){
-				auto result = std::make_shared<Item>(Item());
-				double v = *static_cast<double*>(operands[0]->data);
-				for(unsigned i = 1; i < operands.size(); ++i){
-					v -= *static_cast<double*>(operands[i]->data);
-				}
-				result->write(&v, sizeof(v), ItemTypes::NUMBER);
-				return result;
-			}else
-			if(operands[0]->type == ItemTypes::LIST){
-				if(!doSizesMatch(operands)){
-					cursor->setError("operator '-': lists of mismatching sizes");
-					return std::make_shared<Item>(Item());						
-				}
-				auto firstItem = std::static_pointer_cast<ListType>(operands[0]);
-
-				if(firstItem->list.size() == 0){
-					cursor->setError("operator '-': empty lists");
-					return std::make_shared<Item>(Item());						
-				}
-
-				auto result = std::make_shared<ListType>(ListType( firstItem->list.size() ));
-				for(unsigned i = 0; i < operands.size(); ++i){
-					auto list = std::static_pointer_cast<ListType>(operands[i]);
-					for(unsigned j = 0; j < list->list.size(); ++j){
-						auto from = list->list[j];
-						auto to = result->list[j];
-						double v = *static_cast<double*>(from->data) - *static_cast<double*>(to->data);
-						to->write(&v, sizeof(v), ItemTypes::NUMBER);
-					}
-				}
-				return std::static_pointer_cast<Item>(result);
-			}else{
-				cursor->setError("operator '-': invalid type");
-				return std::make_shared<Item>(Item());				
-			}		
-			
-			return std::make_shared<Item>(Item());				
-		}, {}
-	))); 	
-
-	/*
-	
-		MULTIPLICATION
-
-	*/
-	ctx->registerProperty("*", std::make_shared<FunctionType>(FunctionType([](const std::vector<std::shared_ptr<Item>> &operands, std::shared_ptr<Item> &ctx, Cursor *cursor){
-			if(operands.size() < 2){
-				cursor->setError("operator '*': no operands provided");
-				return std::make_shared<Item>(Item());						
-			}
-			int nilInd = isThereNil(operands);
-			if(nilInd != -1){
-				cursor->setError("operator '*': argument("+std::to_string(nilInd+1)+") is nil");
-				return std::make_shared<Item>(Item());				
-			}
-			if(!doTypesMatch(operands)){
-				cursor->setError("operator '*': mismatching types");
-				return std::make_shared<Item>(Item());
-			}
-
-			if(operands[0]->type == ItemTypes::NUMBER){
-				auto result = std::make_shared<Item>(Item());
-				double v = *static_cast<double*>(operands[0]->data);
-				for(unsigned i = 1; i < operands.size(); ++i){
-					v *= *static_cast<double*>(operands[i]->data);
-				}
-				result->write(&v, sizeof(v), ItemTypes::NUMBER);
-				return result;
-			}else
-			if(operands[0]->type == ItemTypes::LIST){
-				if(!doSizesMatch(operands)){
-					cursor->setError("operator '*': lists of mismatching sizes");
-					return std::make_shared<Item>(Item());						
-				}
-				auto firstItem = std::static_pointer_cast<ListType>(operands[0]);
-
-				if(firstItem->list.size() == 0){
-					cursor->setError("operator '*': empty lists");
-					return std::make_shared<Item>(Item());						
-				}
-
-				auto result = std::make_shared<ListType>(ListType( firstItem->list.size() ));
-				for(unsigned j = 0; j < firstItem->list.size(); ++j){
-					double v = *static_cast<double*>(firstItem->list[j]->data);
-					auto to = result->list[j];
-					to->write(&v, sizeof(v), ItemTypes::NUMBER);
-				}
-
-				for(unsigned i = 1; i < operands.size(); ++i){
-					auto list = std::static_pointer_cast<ListType>(operands[i]);
-					for(unsigned j = 0; j < list->list.size(); ++j){
-						auto from = list->list[j];
-						auto to = result->list[j];
-						double v = *static_cast<double*>(from->data) * *static_cast<double*>(to->data);
-						to->write(&v, sizeof(v), ItemTypes::NUMBER);
-					}
-				}
-				return std::static_pointer_cast<Item>(result);
-			}else{
-				cursor->setError("operator '*': invalid type");
-				return std::make_shared<Item>(Item());				
-			}		
-			
-			return std::make_shared<Item>(Item());				
-		}, {}
-	))); 
-
-	/*
-	
-		DIVISION
-
-	*/
-	ctx->registerProperty("/", std::make_shared<FunctionType>(FunctionType([](const std::vector<std::shared_ptr<Item>> &operands, std::shared_ptr<Item> &ctx, Cursor *cursor){
-			if(operands.size() < 2){
-				cursor->setError("operator '/': no operands provided");
-				return std::make_shared<Item>(Item());						
-			}
-			int nilInd = isThereNil(operands);
-			if(nilInd != -1){
-				cursor->setError("operator '/': argument("+std::to_string(nilInd+1)+") is nil");
-				return std::make_shared<Item>(Item());				
-			}
-			if(!doTypesMatch(operands)){
-				cursor->setError("operator '/': mismatching types");
-				return std::make_shared<Item>(Item());
-			}
-
-			if(operands[0]->type == ItemTypes::NUMBER){
-				auto result = std::make_shared<Item>(Item());
-				double v = *static_cast<double*>(operands[0]->data);
-				for(unsigned i = 1; i < operands.size(); ++i){
-					v /= *static_cast<double*>(operands[i]->data);
-				}
-				result->write(&v, sizeof(v), ItemTypes::NUMBER);
-				return result;
-			}else
-			if(operands[0]->type == ItemTypes::LIST){
-				if(!doSizesMatch(operands)){
-					cursor->setError("operator '/': lists of mismatching sizes");
-					return std::make_shared<Item>(Item());						
-				}
-				auto firstItem = std::static_pointer_cast<ListType>(operands[0]);
-
-				auto result = std::make_shared<ListType>(ListType( firstItem->list.size() ));
-				for(unsigned j = 0; j < firstItem->list.size(); ++j){
-					double v = *static_cast<double*>(firstItem->list[j]->data);
-					auto to = result->list[j];
-					to->write(&v, sizeof(v), ItemTypes::NUMBER);
-				}
-
-				for(unsigned i = 1; i < operands.size(); ++i){
-					auto list = std::static_pointer_cast<ListType>(operands[i]);
-					for(unsigned j = 0; j < list->list.size(); ++j){
-						auto from = list->list[j];
-						auto to = result->list[j];
-						double v = *static_cast<double*>(from->data) / *static_cast<double*>(to->data);
-						to->write(&v, sizeof(v), ItemTypes::NUMBER);
-					}
-				}
-				return std::static_pointer_cast<Item>(result);
-			}else{
-				cursor->setError("operator '/': invalid type");
-				return std::make_shared<Item>(Item());				
-			}		
-			
-			return std::make_shared<Item>(Item());				
-		}, {}
-	))); 	
-
-	/*
-	
-		eq (==)
-
-	*/
-	ctx->registerProperty("eq", std::make_shared<FunctionType>(FunctionType([](const std::vector<std::shared_ptr<Item>> &operands, std::shared_ptr<Item> &ctx, Cursor *cursor){
-			if(operands.size() < 2){
-				cursor->setError("operator 'eq': expects at least 2 operands");
-				return std::make_shared<Item>(Item());						
-			}
-
-			if(!doTypesMatch(operands)){
-				cursor->setError("operator 'eq': mismatching types");
-				return std::make_shared<Item>(Item());
-			}
-
-			if(operands[0]->type == ItemTypes::LIST || operands[0]->type == ItemTypes::PROTO){
-				cursor->setError("operator 'eq': can't compare LIST or PROTOTYPE");
-				return std::make_shared<Item>(Item());
-			}else
-			if(operands[0]->type == ItemTypes::NUMBER){
-				double last = *static_cast<double*>(operands[0]->data);
-				for(int i = 1; i < operands.size(); ++i){
-					if(last !=  *static_cast<double*>(operands[i]->data)){
-						return create(0);
-					}
-				}				
-				return create(1);
-			}else
-			if(operands[0]->type == ItemTypes::STRING){
-				std::string last = static_cast<StringType*>(operands[0].get())->literal;
-				for(int i = 1; i < operands.size(); ++i){
-					if(last != static_cast<StringType*>(operands[i].get())->literal){
-						return create(0);
-					}
-				}				
-				return create(1);
-			}else{
-				cursor->setError("operator 'eq': unsupported operand type");
-				return std::make_shared<Item>(Item());				
-			}
-			
-			return std::make_shared<Item>(Item());				
-		}, {}
-	)));	
-
-	/*
-	
-		neq (==)
-
-	*/
-	ctx->registerProperty("neq", std::make_shared<FunctionType>(FunctionType([](const std::vector<std::shared_ptr<Item>> &operands, std::shared_ptr<Item> &ctx, Cursor *cursor){
-			if(operands.size() < 2){
-				cursor->setError("operator 'neq': expects at least 2 operands");
-				return std::make_shared<Item>(Item());						
-			}
-
-			if(!doTypesMatch(operands)){
-				cursor->setError("operator 'neq': mismatching types");
-				return std::make_shared<Item>(Item());
-			}
-
-			if(operands[0]->type == ItemTypes::LIST || operands[0]->type == ItemTypes::PROTO){
-				cursor->setError("operator 'neq': can't compare LIST or PROTOTYPE");
-				return std::make_shared<Item>(Item());
-			}else
-			if(operands[0]->type == ItemTypes::NUMBER){
-				double last = *static_cast<double*>(operands[0]->data);
-				for(int i = 1; i < operands.size(); ++i){
-					if(last !=  *static_cast<double*>(operands[i]->data)){
-						return create(1);
-					}
-				}				
-				return create(0);
-			}else
-			if(operands[0]->type == ItemTypes::STRING){
-				std::string last = static_cast<StringType*>(operands[0].get())->literal;
-				for(int i = 1; i < operands.size(); ++i){
-					if(last != static_cast<StringType*>(operands[i].get())->literal){
-						return create(1);
-					}
-				}				
-				return create(0);
-			}else{
-				cursor->setError("operator 'neq': unsupported operand type");
-				return std::make_shared<Item>(Item());				
-			}
-			
-			return std::make_shared<Item>(Item());				
-		}, {}
-	)));	
-
-	/*
-	
-		gt (>)
-
-	*/
-	ctx->registerProperty("gt", std::make_shared<FunctionType>(FunctionType([](const std::vector<std::shared_ptr<Item>> &operands, std::shared_ptr<Item> &ctx, Cursor *cursor){
-			if(operands.size() < 2){
-				cursor->setError("operator 'gt': expects at least 2 operands");
-				return std::make_shared<Item>(Item());						
-			}
-
-			if(!doTypesMatch(operands)){
-				cursor->setError("operator 'gt': mismatching types");
-				return std::make_shared<Item>(Item());
-			}
-
-			if(operands[0]->type == ItemTypes::LIST || operands[0]->type == ItemTypes::PROTO){
-				int last = *static_cast<double*>(operands[0]->getProperty("length")->data);
-				for(int i = 1; i < operands.size(); ++i){
-					if(last <= *static_cast<double*>(operands[i]->getProperty("length")->data)){
-						return create(0);
-					}
-					last = *static_cast<double*>(operands[i]->getProperty("length")->data);
-				}				
-				return create(1);
-			}else		
-			if(operands[0]->type == ItemTypes::NUMBER){
-				double last = *static_cast<double*>(operands[0]->data);
-				for(int i = 1; i < operands.size(); ++i){
-					if(last <=  *static_cast<double*>(operands[i]->data)){
-						return create(0);
-					}
-					last = *static_cast<double*>(operands[i]->data);
-				}				
-				return create(1);
-			}else
-			if(operands[0]->type == ItemTypes::STRING){
-				std::string last = static_cast<StringType*>(operands[0].get())->literal;
-				for(int i = 1; i < operands.size(); ++i){
-					if(last.size() <= static_cast<StringType*>(operands[i].get())->literal.size()){
-						return create(0);
-					}
-					last = static_cast<StringType*>(operands[i].get())->literal;
-				}				
-				return create(1);
-			}else{
-				cursor->setError("operator 'gt': unsupported operand type");
-				return std::make_shared<Item>(Item());				
-			}
-			
-			return std::make_shared<Item>(Item());				
-		}, {}
-	)));		
-
-	/*
-	
-		gte (>=)
-
-	*/
-	ctx->registerProperty("gte", std::make_shared<FunctionType>(FunctionType([](const std::vector<std::shared_ptr<Item>> &operands, std::shared_ptr<Item> &ctx, Cursor *cursor){
-			if(operands.size() < 2){
-				cursor->setError("operator 'gte': expects at least 2 operands");
-				return std::make_shared<Item>(Item());						
-			}
-
-			if(!doTypesMatch(operands)){
-				cursor->setError("operator 'gte': mismatching types");
-				return std::make_shared<Item>(Item());
-			}
-
-			if(operands[0]->type == ItemTypes::LIST || operands[0]->type == ItemTypes::PROTO){
-				int last = *static_cast<double*>(operands[0]->getProperty("length")->data);
-				for(int i = 1; i < operands.size(); ++i){
-					if(last < *static_cast<double*>(operands[i]->getProperty("length")->data)){
-						return create(0);
-					}
-					last = *static_cast<double*>(operands[i]->getProperty("length")->data);
-				}				
-				return create(1);
-			}else		
-			if(operands[0]->type == ItemTypes::NUMBER){
-				double last = *static_cast<double*>(operands[0]->data);
-				for(int i = 1; i < operands.size(); ++i){
-					if(last <  *static_cast<double*>(operands[i]->data)){
-						return create(0);
-					}
-					last = *static_cast<double*>(operands[i]->data);
-				}				
-				return create(1);
-			}else
-			if(operands[0]->type == ItemTypes::STRING){
-				std::string last = static_cast<StringType*>(operands[0].get())->literal;
-				for(int i = 1; i < operands.size(); ++i){
-					if(last.size() < static_cast<StringType*>(operands[i].get())->literal.size()){
-						return create(0);
-					}
-					last = static_cast<StringType*>(operands[i].get())->literal;
-				}				
-				return create(1);
-			}else{
-				cursor->setError("operator 'gte': unsupported operand type");
-				return std::make_shared<Item>(Item());				
-			}
-			
-			return std::make_shared<Item>(Item());				
-		}, {}
-	)));	
-
-	/*
-	
-		lt (<)
-
-	*/
-	ctx->registerProperty("lt", std::make_shared<FunctionType>(FunctionType([](const std::vector<std::shared_ptr<Item>> &operands, std::shared_ptr<Item> &ctx, Cursor *cursor){
-			if(operands.size() < 2){
-				cursor->setError("operator 'lt': expects at least 2 operands");
-				return std::make_shared<Item>(Item());						
-			}
-
-			if(!doTypesMatch(operands)){
-				cursor->setError("operator 'lt': mismatching types");
-				return std::make_shared<Item>(Item());
-			}
-
-			if(operands[0]->type == ItemTypes::LIST || operands[0]->type == ItemTypes::PROTO){
-				int last = *static_cast<double*>(operands[0]->getProperty("length")->data);
-				for(int i = 1; i < operands.size(); ++i){
-					if(last >= *static_cast<double*>(operands[i]->getProperty("length")->data)){
-						return create(0);
-					}
-					last = *static_cast<double*>(operands[i]->getProperty("length")->data);
-				}				
-				return create(1);
-			}else		
-			if(operands[0]->type == ItemTypes::NUMBER){
-				double last = *static_cast<double*>(operands[0]->data);
-				for(int i = 1; i < operands.size(); ++i){
-					if(last >=  *static_cast<double*>(operands[i]->data)){
-						return create(0);
-					}
-					last = *static_cast<double*>(operands[i]->data);
-				}				
-				return create(1);
-			}else
-			if(operands[0]->type == ItemTypes::STRING){
-				std::string last = static_cast<StringType*>(operands[0].get())->literal;
-				for(int i = 1; i < operands.size(); ++i){
-					if(last.size() >= static_cast<StringType*>(operands[i].get())->literal.size()){
-						return create(0);
-					}
-					last = static_cast<StringType*>(operands[i].get())->literal;
-				}				
-				return create(1);
-			}else{
-				cursor->setError("operator 'lt': unsupported operand type");
-				return std::make_shared<Item>(Item());				
-			}
-			
-			return std::make_shared<Item>(Item());				
-		}, {}
-	)));		
-
-	/*
-	
-		lte (<=)
-
-	*/
-	ctx->registerProperty("lte", std::make_shared<FunctionType>(FunctionType([](const std::vector<std::shared_ptr<Item>> &operands, std::shared_ptr<Item> &ctx, Cursor *cursor){
-			if(operands.size() < 2){
-				cursor->setError("operator 'lte': expects at least 2 operands");
-				return std::make_shared<Item>(Item());						
-			}
-
-			if(!doTypesMatch(operands)){
-				cursor->setError("operator 'lte': mismatching types");
-				return std::make_shared<Item>(Item());
-			}
-
-			if(operands[0]->type == ItemTypes::LIST || operands[0]->type == ItemTypes::PROTO){
-				int last = *static_cast<double*>(operands[0]->getProperty("length")->data);
-				for(int i = 1; i < operands.size(); ++i){
-					if(last > *static_cast<double*>(operands[i]->getProperty("length")->data)){
-						return create(0);
-					}
-					last = *static_cast<double*>(operands[i]->getProperty("length")->data);
-				}				
-				return create(1);
-			}else		
-			if(operands[0]->type == ItemTypes::NUMBER){
-				double last = *static_cast<double*>(operands[0]->data);
-				for(int i = 1; i < operands.size(); ++i){
-					if(last >  *static_cast<double*>(operands[i]->data)){
-						return create(0);
-					}
-					last = *static_cast<double*>(operands[i]->data);
-				}				
-				return create(1);
-			}else
-			if(operands[0]->type == ItemTypes::STRING){
-				std::string last = static_cast<StringType*>(operands[0].get())->literal;
-				for(int i = 1; i < operands.size(); ++i){
-					if(last.size() > static_cast<StringType*>(operands[i].get())->literal.size()){
-						return create(0);
-					}
-					last = static_cast<StringType*>(operands[i].get())->literal;
-				}				
-				return create(1);
-			}else{
-				cursor->setError("operator 'lte': unsupported operand type");
-				return std::make_shared<Item>(Item());				
-			}
-			
-			return std::make_shared<Item>(Item());				
-		}, {}
-	)));	
-
-	/*
-	
-		..
-
-	*/
-	ctx->registerProperty("..", std::make_shared<FunctionType>(FunctionType([](const std::vector<std::shared_ptr<Item>> &operands, std::shared_ptr<Item> &ctx, Cursor *cursor){
-			if(operands.size() < 2){
-				cursor->setError("operator '..': expects at least 2 operands");
-				return std::make_shared<Item>(Item());						
-			}
-
-			if(operands.size() > 3){
-				cursor->setError("operator '..': expects no more than 3 operands");
-				return std::make_shared<Item>(Item());						
-			}			
-
-			// 
-
-			if(!doTypesMatch(operands)){
-				cursor->setError("operator '..': mismatching types");
-				return std::make_shared<Item>(Item());
-			}
-
-			if(operands[0]->type == ItemTypes::NUMBER){
-				double s = *static_cast<double*>(operands[0]->data);
-				double e = *static_cast<double*>(operands[1]->data);
-				auto result = createList(0);
-				if(s == e){
-					return std::static_pointer_cast<Item>(result);
-				}
-				bool inc = e > s;
-				double step = Tools::positive(operands.size() == 3 ? *static_cast<double*>(operands[2]->data) : 1);
-
-				for(double i = s; (inc ? i <= e : i >= e); i += (inc ? step : -step)){
-					result->add(create(i));
-				}
-
-				return std::static_pointer_cast<Item>(result);
-			}else{
-				cursor->setError("operator '..': unsupported operand type");
-				return std::make_shared<Item>(Item());				
-			}
-			
-			return std::make_shared<Item>(Item());				
-		}, {}
-	)));	
-
-	/*
-	
-		splice
-
-	*/
-	ctx->registerProperty("splice", std::make_shared<FunctionType>(FunctionType([](const std::vector<std::shared_ptr<Item>> &operands, std::shared_ptr<Item> &ctx, Cursor *cursor){
-			
-			if(operands.size() < 2){
-				cursor->setError("operator 'splice': expects at least 2 operands");
-				return std::make_shared<Item>(Item());						
-			}			
-
-			if(!doTypesMatch(operands)){
-				cursor->setError("operator 'splice': mismatching types");
-				return std::make_shared<Item>(Item());
-			}		
-
-			if(operands[0]->type == ItemTypes::LIST){
-				auto list = createList(0);
-				for(int i = 0; i < operands.size(); ++i){
-					auto operand = std::static_pointer_cast<ListType>(operands[i]);
-					for(int j = 0; j < operand->list.size(); ++j){
-						list->add(operand->list[j]);
-					}
-				}
-				return std::static_pointer_cast<Item>(list);
-			}else
-			if(operands[0]->type == ItemTypes::PROTO){
-				auto proto = createProto();
-				for(int i = 0; i < operands.size(); ++i){
-					auto operand = std::static_pointer_cast<ProtoType>(operands[i]);
-					for(auto &it : operand->members){
-						std::string name = it.first;
-						proto->add(name, it.second, ctx, cursor);
-					}
-				}
-				return std::static_pointer_cast<Item>(proto);
-			}else{
-				cursor->setError("operator 'splice': unsupported operand type");
-				return std::make_shared<Item>(Item());					
-			}
-
-		}, {}
-
-	)));	
-
-	/*
-	
-		l-flatten
-
-	*/
-	ctx->registerProperty("l-flatten", std::make_shared<FunctionType>(FunctionType([](const std::vector<std::shared_ptr<Item>> &operands, std::shared_ptr<Item> &ctx, Cursor *cursor){
-			
-			if(operands.size() < 1){
-				cursor->setError("operator 'l-flatten': expects at least 1 operand");
-				return std::make_shared<Item>(Item());						
-			}			
-
-			if(!doTypesMatch(operands)){
-				cursor->setError("operator 'l-flatten': mismatching types");
-				return std::make_shared<Item>(Item());
-			}
-
-			if(operands[0]->type != ItemTypes::LIST){
-				cursor->setError("operator 'l-flatten': expects only lists");
-				return std::make_shared<Item>(Item());
-			}
-			
-			auto out = createList(0);
-			std::function<void(const std::shared_ptr<Item> &obj)> recursive = [&](const std::shared_ptr<Item> &obj){
-				if(obj->type == ItemTypes::LIST){
-					auto list = std::static_pointer_cast<ListType>(obj);
-					for(int i = 0; i < list->list.size(); ++i){
-						recursive(list->list[i]);
-					}
-				}else{
-					out->add(obj);
-				}
-			};
-
-			for(int i = 0; i < operands.size(); ++i){
-				recursive(operands[i]);
-			}
-
-
-			return std::static_pointer_cast<Item>(out);
-
-
-		}, {}
-
-	)));
-
-
-
-	/*
-	
-		++
-
-	*/
-	ctx->registerProperty("++", std::make_shared<FunctionType>(FunctionType([](const std::vector<std::shared_ptr<Item>> &operands, std::shared_ptr<Item> &ctx, Cursor *cursor){
-			
-			if(operands.size() < 1){
-				cursor->setError("operator '++': expects at least 1 operand");
-				return std::make_shared<Item>(Item());						
-			}			
-
-			if(!doTypesMatch(operands)){
-				cursor->setError("operator '++': mismatching types");
-				return std::make_shared<Item>(Item());
-			}
-
-			if(operands[0]->type != ItemTypes::NUMBER){
-				cursor->setError("operator '++': expects only lists");
-				return std::make_shared<Item>(Item());
-			}
-
-			for(int i = 0; i < operands.size(); ++i){
-				auto item = operands[i];
-				double v = *static_cast<double*>(item->data);
-				v += 1;
-				item->write(&v, sizeof(v), ItemTypes::NUMBER);
-			}
-
-			return operands[0];
-		}, {}
-	)));	
-
-	/*
-	
-		--
-
-	*/
-	ctx->registerProperty("--", std::make_shared<FunctionType>(FunctionType([](const std::vector<std::shared_ptr<Item>> &operands, std::shared_ptr<Item> &ctx, Cursor *cursor){
-			
-			if(operands.size() < 1){
-				cursor->setError("operator '--': expects at least 1 operand");
-				return std::make_shared<Item>(Item());						
-			}			
-
-			if(!doTypesMatch(operands)){
-				cursor->setError("operator '--': mismatching types");
-				return std::make_shared<Item>(Item());
-			}
-
-			if(operands[0]->type != ItemTypes::NUMBER){
-				cursor->setError("operator '--': expects only lists");
-				return std::make_shared<Item>(Item());
-			}
-
-			for(int i = 0; i < operands.size(); ++i){
-				auto item = operands[i];
-				double v = *static_cast<double*>(item->data);
-				v -= 1;
-				item->write(&v, sizeof(v), ItemTypes::NUMBER);
-			}
-
-			return operands[0];
-		}, {}
-	)));
-
-	/*
-	
-		noop
-
-	*/
-	ctx->registerProperty("noop", std::make_shared<FunctionType>(FunctionType([](const std::vector<std::shared_ptr<Item>> &operands, std::shared_ptr<Item> &ctx, Cursor *cursor){
-			
-			return std::make_shared<Item>(Item());
-
-		}, {}
-	)));		
-
-	/*
-
-		join
-	
-	*/
-	ctx->registerProperty("s-join", std::make_shared<FunctionType>(FunctionType([](const std::vector<std::shared_ptr<Item>> &operands, std::shared_ptr<Item> &ctx, Cursor *cursor){
-			if(operands.size() < 2){
-				cursor->setError("operator 'join': expects at least 2 operands");
-				return std::make_shared<Item>(Item());						
-			}
-
-			std::string output;
-			for(int i = 0; i < operands.size(); ++i){
-				output += operands[i]->str(false);
-			}
-
-			return std::static_pointer_cast<Item>(CV::createString(output));
-			
-		}, {}
-	)));	
-
-
-
-	/*
-
-		eq-ref
-	
-	*/
-	ctx->registerProperty("eq-ref", std::make_shared<FunctionType>(FunctionType([](const std::vector<std::shared_ptr<Item>> &operands, std::shared_ptr<Item> &ctx, Cursor *cursor){
-			if(operands.size() < 2){
-				cursor->setError("operator 'eq-ref': expects at least 2 operands");
-				return std::make_shared<Item>(Item());						
-			}
-
-			void *last = operands[0].get();
-			for(int i = 1; i < operands.size(); ++i){
-				if(static_cast<Item*>(last)->type != operands[i]->type || operands[i].get() != last){
-					return create(0);
-				}
-			}
-
-			return create(1);
-			
-		}, {}
-	)));	
-
-
-	/*
-
-		quit
-	
-	*/
-	ctx->registerProperty("quit", std::make_shared<FunctionType>(FunctionType([](const std::vector<std::shared_ptr<Item>> &operands, std::shared_ptr<Item> &ctx, Cursor *cursor){
-			std::exit(0);
-			return std::make_shared<Item>(Item());				
-		}, {}
-	)));		
-
+void CV::Function::set(const std::vector<std::string> &params, const std::function<std::shared_ptr<CV::Item> (const std::vector<std::shared_ptr<CV::Item>> &params, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx)> &fn, bool variadic){
+    this->params = params;
+    this->type = CV::ItemTypes::FUNCTION;
+    this->fn = fn;
+    this->binary = true;
+    this->variadic = variadic;
+}
+
+std::shared_ptr<CV::Item> CV::Context::get(const std::string &name){
+    auto it = vars.find(name);
+    if(it == vars.end()){
+        return this->head ? this->head->get(name) : std::shared_ptr<CV::Item>(NULL);
+    }
+    return it->second;
+}
+
+std::shared_ptr<CV::Item> CV::Context::copy(bool deep) const {
+
+    auto item = std::make_shared<CV::Context>(CV::Context());
+    item->type = this->type;  
+    item->head = this->head;    
+
+    for(auto &it : this->vars){
+        item->vars[it.first] = deep ? it.second->copy() : it.second;
+    }        
+
+    return item;
+}
+
+void CV::Context::setTop(std::shared_ptr<CV::Context> &nctx){
+    auto ctop = this->get("top");
+    if(ctop){
+        return;
+    }
+    auto ctx = std::static_pointer_cast<CV::Item>(nctx);
+    this->set("top", ctx);
+}
+
+CV::ItemContextPair CV::Context::getWithContext(const std::string &name){
+    auto it = vars.find(name);
+    if(it == vars.end()){
+        return this->head ? this->head->getWithContext(name) : ItemContextPair();
+    }
+    return ItemContextPair(it->second.get(), this, it->first);
+}    
+
+CV::ItemContextPair CV::Context::getWithContext(std::shared_ptr<CV::Item> &item){
+    for(auto &it : this->vars){
+        if(it.second.get() == item.get()){
+            return ItemContextPair(it.second.get(), this, it.first);
+        }
+    }
+    return this->head ? this->head->getWithContext(item) : ItemContextPair();
 }        
+
+std::shared_ptr<CV::Item> CV::Context::set(const std::string &name, const std::shared_ptr<CV::Item> &item){
+    auto v = get(name);
+    this->vars[name] = item;
+    return item;
+}
+
+CV::Context::Context(){
+    this->type = CV::ItemTypes::CONTEXT;
+    this->head = NULL;
+}
+
+CV::Context::Context(std::shared_ptr<CV::Context> &ctx){
+    this->type = CV::ItemTypes::CONTEXT;        
+    this->head = ctx;
+}    
+
+
+bool CV::Context::clear(bool deep){
+    if(this->type == CV::ItemTypes::NIL){
+        return false;
+    }
+
+    return true;
+}
+
+void CV::Context::debug(){
+    std::cout << "\n\n==================CONTEXT DEBUG==================" << std::endl;
+    size_t maxVarName = 0;
+    auto getSpace = [&](int w){
+        std::string space = "";
+        int max = std::max(maxVarName, (size_t)10) - w;
+        for(int i = 0; i < max; ++i) space += " ";
+        return space;
+    };
+    for(auto &it : this->vars){
+        maxVarName = std::max(maxVarName, it.first.length());
+    }
+
+    for(auto &it : this->vars){
+        std::cout << it.first << getSpace(it.first.length()) << " -> " << getSpace(0) << CV::ItemToText(it.second.get()) << std::endl;
+    }
+}
+
+std::string CV::ItemToText(CV::Item *item){
+    switch(item->type){
+        default:
+        case CV::ItemTypes::NIL: {
+            return "nil";
+        };
+        case CV::ItemTypes::INTERRUPT: {
+            auto interrupt = static_cast<CV::Interrupt*>(item);
+            return "[INT-"+CV::InterruptTypes::str(interrupt->intType)+(interrupt->hasPayload() ? " "+CV::ItemToText(interrupt->getPayload().get())+"" : "")+"]";
+        };                
+        case CV::ItemTypes::NUMBER: {
+            auto r = Tools::removeTrailingZeros(static_cast<CV::Number*>(item)->n);
+            return r;
+        };
+        case CV::ItemTypes::STRING: {
+            return "'"+static_cast<CV::String*>(item)->data+"'";
+        };        
+        case CV::ItemTypes::FUNCTION: {
+            auto fn = static_cast<CV::Function*>(item);        
+            return "[fn ["+(fn->variadic ? "..." : Tools::compileList(fn->params))+"] ["+(fn->binary ? "BINARY" : fn->body )+"]]";
+        };
+        case CV::ItemTypes::CONTEXT: {
+            std::string output = "[ct ";
+            auto proto = static_cast<CV::Context*>(item);
+            int n = proto->vars.size();
+            int c = 0;
+            for(auto &it : proto->vars){
+                auto name = it.first;
+                auto item = it.second;
+                output += CV::ItemToText(item.get())+CV::ModifierTypes::str(CV::ModifierTypes::NAMER)+name;
+                if(c < n-1){
+                    output += " ";
+                }
+                ++c;                
+            }
+            return output + "]";            
+        };
+        case CV::ItemTypes::LIST: {
+            std::string output = "[";
+            auto list = static_cast<CV::List*>(item);
+            for(int i = 0; i < list->data.size(); ++i){
+                auto item = list->get(i);
+                output += CV::ItemToText(item.get());
+                if(i < list->data.size()-1){
+                    output += " ";
+                }
+            }
+            return output + "]";
+        };         
+    }
+}
+
+
+
+static std::vector<std::string> parseTokens(std::string input, char sep, CV::Cursor *cursor){
+    std::vector<std::string> tokens; 
+    std::string buffer;
+    int open = 0;
+    bool onString = false;
+    bool onComment = false;
+    int leftBrackets = 0;
+    int rightBrackets = 0;
+    int cline = 0;
+    int quotes = 0;
+    // Count outer brackets
+    for(int i = 0; i < input.size(); ++i){
+        char c = input[i];
+        if(c == '\n'){
+            ++cline;
+        }else
+        if(c == '\''){
+            onString = !onString;
+            if(onString){
+                ++quotes;
+            }else{
+                --quotes;
+            }
+        }else
+        if(c == '[' && !onString){
+            if(open == 0) ++leftBrackets;
+            ++open;
+        }else
+        if(c == ']' && !onString){
+            if(open == 1) ++rightBrackets;
+            --open;
+        }
+    }
+    onString = false;
+    open = 0;
+    // Throw mismatching brackets
+    if(leftBrackets != rightBrackets){
+        cursor->setError("Mismatching brackets", cline);
+        return {};
+    }
+    // Throw mismatching quotes
+    if(quotes != 0){
+        cursor->setError("Mismatching quotes", cline);
+        return {};
+    }
+    cline = 0;
+
+
+    // Add missing brackets for simple statements
+    if(((leftBrackets+rightBrackets)/2 > 1) || (leftBrackets+rightBrackets) == 0 || input[0] != '[' || input[input.length()-1] != ']'){
+        input = "[" + input + "]";
+    }
+    // For debugging:
+    // std::cout << "To Parse '" << input << "' " << leftBrackets+rightBrackets << std::endl;
+    // Parse
+    for(int i = 0; i < input.size(); ++i){
+        char c = input[i];
+        if(c == '\'' && !onComment){
+            onString = !onString;
+            buffer += c;
+        }else        
+        if(c == '#' && !onString){
+            onComment = !onComment;
+        }else
+        if(c == '\n' && !onString){
+            onComment = false;
+            continue;
+        }else
+        if(onComment){
+            continue;
+        }else
+        if(c == '[' && !onString){
+            if(open > 0){
+                buffer += c;
+            }
+            ++open;
+        }else
+        if(c == ']' && !onString){
+            if(open > 1){
+                buffer += c;
+            }            
+            --open;
+            if(open == 1 || i == input.size()-1 && buffer.length() > 0){
+                if(buffer != ""){
+                    tokens.push_back(buffer);
+                }
+                buffer = "";
+            }
+        }else
+        if(i == input.size()-1){
+            if(buffer != ""){
+                tokens.push_back(buffer+c);
+            }
+            buffer = "";
+        }else
+        if(c == sep && open == 1 && !onString){
+            if(buffer != ""){
+                tokens.push_back(buffer);
+            }
+            buffer = "";
+        }else{
+            buffer += c;
+        }
+    }
+
+    return tokens;
+};
+
+CV::Token parseTokenWModifier(const std::string &_input){
+    std::string buffer;
+    CV::Token token;
+    std::string input = _input;
+    if(input[0] == '[' && input[input.size()-1] == ']'){
+        input = input.substr(1, input.length() - 2);
+    }
+    for(int i = 0; i < input.length(); ++i){
+        char c = input[input.length()-1 - i];
+        auto t = CV::ModifierTypes::getToken(c);
+        if(t != CV::ModifierTypes::UNDEFINED){
+                std::reverse(buffer.begin(), buffer.end());            
+                token.modifiers.push_back(CV::ModifierPair(t, buffer));
+                buffer = "";   
+        }else
+        if(t == CV::ModifierTypes::UNDEFINED && i == input.length()-1){
+            buffer += c;
+            token.first = buffer;
+            std::reverse(token.first.begin(), token.first.end());
+            buffer = "";
+        }else{
+            buffer += c;
+        }
+    }
+    std::reverse(token.modifiers.begin(), token.modifiers.end());
+    return token;
+}
+
+std::vector<CV::Token> buildTokens(const std::vector<std::string> &literals, CV::Cursor *cursor){
+
+    std::vector<CV::Token> tokens;
+    for(int i = 0; i < literals.size(); ++i){
+        auto innLiterals = parseTokens(literals[i], ' ', cursor);
+        // Complex tokens (It's usually tokens that actually are complete statements) need to be evaluated further        
+        if(innLiterals.size() > 1){
+            CV::Token token;
+            token.first = literals[i];
+            tokens.push_back(token);  
+        // Simple can be inserted right away
+        }else{
+            auto token = parseTokenWModifier(literals[i]);
+            if(token.first == "" && i > 0){ // This means this modifier belongs to the token before this one
+                auto &prev = tokens[tokens.size()-1];
+                for(int j = 0; j < token.modifiers.size(); ++j){
+                    prev.modifiers.push_back(token.modifiers[j]);
+                }
+                continue;
+            }else{
+                tokens.push_back(token);
+            }
+        }
+    }
+
+    return tokens;
+};
+
+static std::vector<CV::Token> compileTokens(const std::vector<CV::Token> &tokens, int from, int amount){
+	std::vector<CV::Token> compiled;
+	for(int i = 0; i < amount; ++i){
+        compiled.push_back(tokens[from + i]);
+	}
+	return compiled;
+}
+
+bool CV::FunctionConstraints::test(List *list, std::string &errormsg){
+    std::vector<std::shared_ptr<CV::Item>> items;
+    for(int i = 0; i < list->data.size(); ++i){
+        items.push_back(list->data[i]);
+    }
+    return this->test(items, errormsg);
+}
+
+bool CV::FunctionConstraints::test(const std::vector<std::shared_ptr<CV::Item>> &items, std::string &errormsg){
+    if(!enabled || items.size() == 0){
+        return true;
+    }
+
+    if(useMaxParams && items.size() > maxParams){
+        errormsg = "Provided ("+std::to_string(items.size())+") argument(s). Expected no more than ("+std::to_string(maxParams)+") arguments.";
+        return false;
+    }
+
+    if(useMinParams && items.size() < minParams){
+        errormsg = "Provided ("+std::to_string(items.size())+") argument(s). Expected no less than ("+std::to_string(minParams)+") arguments.";
+        return false;
+    } 
+
+
+    if(!allowNil){
+        for(int i = 1; i < items.size(); ++i){
+            if(items[i]->type == CV::ItemTypes::NIL){
+                errormsg = "Provided nil value.";
+                return false;
+            }
+        }        
+    }
+
+
+    if(!allowMisMatchingTypes){
+        uint8_t firstType = CV::ItemTypes::NIL;
+        for(int i = 0; i < items.size(); ++i){
+            if(items[i]->type != CV::ItemTypes::NIL && i < items.size()-1){
+                firstType = items[i]->type;
+            }
+        }        
+        for(int i = 1; i < items.size(); ++i){
+            if(items[i]->type != firstType && items[i]->type != CV::ItemTypes::NIL){
+                errormsg = "Provided arguments with mismatching types.";
+                return false;
+            }
+        }
+    }
+
+    if(expectedTypeAt.size() > 0){
+        for(int i = 0; i < items.size(); ++i){
+            if(expectedTypeAt.count(i) > 0){
+                if(items[i]->type != expectedTypeAt[i]){
+                    errormsg = "Provided '"+CV::ItemTypes::str(items[i]->type)+"' at position "+std::to_string(i)+". Expected type is '"+CV::ItemTypes::str(expectedTypeAt[i])+"'.";
+                    return false;
+                }
+            }
+        }
+    }
+
+    if(expectedTypes.size() > 0){
+        for(int i = 0; i < items.size(); ++i){
+            bool f = false;
+            for(int j = 0; j < this->expectedTypes.size(); ++j){
+                auto exType = this->expectedTypes[j];
+                if(items[i]->type == exType){
+                    f = true;
+                    break;
+                }
+
+            }
+            if(!f){
+                std::string exTypes = "";
+                for(int i = 0; i < expectedTypes.size(); ++i){
+                    exTypes += "'"+CV::ItemTypes::str(expectedTypes[i])+"'";
+                    if(i < expectedTypes.size()-1){
+                        exTypes += ", ";
+                    }
+                }
+                errormsg = "Provided '"+CV::ItemTypes::str(items[i]->type)+"'. Expected types are only "+exTypes+".";
+                return false;
+            }
+
+        }        
+    }    
+
+    return true;
+}
+
+
+static std::shared_ptr<CV::Item> eval(const std::vector<CV::Token> &tokens, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx);
+
+
+static std::shared_ptr<CV::Item> processPreInterpretModifiers(std::shared_ptr<CV::Item> &item, std::vector<CV::ModifierPair> modifiers, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx, CV::ModifierEffect &effects){
+    effects.reset();
+    for(int i = 0; i < modifiers.size(); ++i){
+        auto &mod = modifiers[i];
+        switch(mod.type){
+            case CV::ModifierTypes::TRAIT: {
+                auto name = mod.subject;
+                if(name.size() == 0){
+                    cursor->setError("'"+CV::ModifierTypes::str(CV::ModifierTypes::SCOPE)+"': Trait modifier requires a name");
+                    return item;                    
+                }
+                // ANY_NUMBER type trait
+                if(CV::Tools::isNumber(name) && item->hasTrait(CV::TraitType::ANY_NUMBER)){
+                    return item->runTrait(CV::TraitType::ANY_NUMBER, name, cursor, ctx);    
+                }else
+                // ANY_STRING type trait
+                if(!CV::Tools::isNumber(name) && item->hasTrait(CV::TraitType::ANY_STRING)){
+                    return item->runTrait(CV::TraitType::ANY_STRING, name, cursor, ctx);    
+                }else                
+                // Regular trait
+                if(!item->hasTrait(name)){
+                    cursor->setError("'"+CV::ModifierTypes::str(CV::ModifierTypes::SCOPE)+"': Trait '"+name+"' doesn't belong to '"+CV::ItemTypes::str(item->type)+"' type");
+                    return item;                      
+                }
+                item = item->runTrait(name, "", cursor, ctx);
+            } break;            
+            case CV::ModifierTypes::LINKER: {
+
+            } break;
+            case CV::ModifierTypes::NAMER: {
+                effects.named = mod.subject;
+                ctx->set(mod.subject, item);
+            } break;        
+            case CV::ModifierTypes::SCOPE: {
+                if(item->type != CV::ItemTypes::CONTEXT){
+                    cursor->setError("'"+CV::ModifierTypes::str(CV::ModifierTypes::SCOPE)+"': Scope modifier can only be used on contexts.");
+                    return item;
+                }
+                if(mod.subject.size() > 0){
+                    auto subsequent = std::static_pointer_cast<CV::Context>(item)->get(mod.subject);
+                    if(!subsequent){
+                        cursor->setError("'"+CV::ModifierTypes::str(CV::ModifierTypes::SCOPE)+"':"+mod.subject+"': Undefined symbol or imperative.");
+                        return item;
+                    }
+                    item = subsequent;
+                }else{
+                    effects.toCtx = std::static_pointer_cast<CV::Context>(item);
+                    effects.ctxSwitch = true;
+                }
+            } break;   
+            case CV::ModifierTypes::EXPAND: {
+                if(i == modifiers.size()-1){
+                    effects.expand = true;
+                }else{
+                    cursor->setError("'"+CV::ModifierTypes::str(CV::ModifierTypes::EXPAND)+"': Expand modifier can only be used at the end of a modifier chain.");
+                    return item;
+                }
+            } break;                                   
+        }
+    }
+    return item;
+}
+
+// Interpret from STRING
+std::shared_ptr<CV::Item> CV::interpret(const std::string &input, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+    auto literals = parseTokens(input, ' ', cursor);
+    // printList(literals);
+    if(cursor->error){
+        return std::make_shared<CV::Item>(CV::Item());
+    }    
+    auto tokens = buildTokens(literals, cursor);
+    return eval(tokens, cursor, ctx);
+}
+
+// Interpret from single TOKEN without dropping modifiers. Should be only used by `eval`
+std::shared_ptr<CV::Item> CV::interpret(const CV::Token &token, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+    auto literals = parseTokens(token.first, ' ', cursor);
+    if(cursor->error){
+        return std::make_shared<CV::Item>(CV::Item());
+    }
+    auto tokens = buildTokens(literals, cursor);
+    return eval(tokens, cursor, ctx);
+}
+
+static std::shared_ptr<CV::Item> processIteration(const std::vector<CV::Token> &tokens, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx);
+static void processPostInterpretExpandListOrContext(std::shared_ptr<CV::Item> &solved, CV::ModifierEffect &effects, std::vector<std::shared_ptr<CV::Item>> &items, CV::Cursor *cursor, const std::shared_ptr<CV::Context> &ctx);
+
+
+static std::shared_ptr<CV::Item> runFunction(std::shared_ptr<CV::Function> &fn, const std::vector<std::shared_ptr<CV::Item>> items, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx);
+static std::shared_ptr<CV::Item> runFunction(std::shared_ptr<CV::Function> &fn, const std::vector<CV::Token> &tokens, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx);
+
+static std::shared_ptr<CV::Item> runFunction(std::shared_ptr<CV::Function> &fn, const std::vector<CV::Token> &tokens, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+    std::vector<std::shared_ptr<CV::Item>> items;
+    for(int i = 0; i < tokens.size(); ++i){
+        CV::ModifierEffect effects;
+        auto interp = interpret(tokens[i], cursor, ctx);
+        auto solved = processPreInterpretModifiers(interp, tokens[i].modifiers, cursor, ctx, effects);                     
+        if(cursor->error){
+            return std::make_shared<CV::Item>(CV::Item());
+        }
+        processPostInterpretExpandListOrContext(solved, effects, items, cursor, std::shared_ptr<CV::Context>((CV::Context*)NULL));
+        if(cursor->error){
+            return std::make_shared<CV::Item>(CV::Item());
+        } 
+    }
+    return runFunction(fn, items, cursor, ctx);
+}
+
+
+static std::shared_ptr<CV::Item> runFunction(std::shared_ptr<CV::Function> &fn, const std::vector<std::shared_ptr<CV::Item>> items, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+    if(fn->binary){
+        return fn->fn(items, cursor, ctx);  
+    }else{
+        auto tctx = std::make_shared<CV::Context>(CV::Context(ctx));
+        tctx->setTop(ctx);
+        if(fn->variadic){
+            auto list = std::make_shared<CV::List>(CV::List(items, false));
+            tctx->set("...", std::static_pointer_cast<CV::Item>(list));
+        }else{
+            if(items.size() != fn->params.size()){
+                cursor->setError("'"+CV::ItemToText(fn.get())+"': Expects exactly "+(std::to_string(fn->params.size()))+" arguments");
+                return std::make_shared<CV::Item>(CV::Item());
+            }
+            for(int i = 0; i < items.size(); ++i){
+                tctx->set(fn->params[i], items[i]);
+            }
+        }
+        return interpret(fn->body, cursor, tctx);
+    }
+}
+
+static void processPostInterpretExpandListOrContext(std::shared_ptr<CV::Item> &solved, CV::ModifierEffect &effects, std::vector<std::shared_ptr<CV::Item>> &items, CV::Cursor *cursor, const std::shared_ptr<CV::Context> &ctx){
+    if(effects.expand && solved->type == CV::ItemTypes::LIST){
+        auto list = std::static_pointer_cast<CV::List>(solved);
+        for(int j = 0; j < list->data.size(); ++j){
+            items.push_back(list->get(j));
+        }
+    }else
+    if(effects.expand && solved->type == CV::ItemTypes::CONTEXT){
+        if(effects.expand && solved->type == CV::ItemTypes::CONTEXT){
+            auto proto = std::static_pointer_cast<CV::Context>(solved);
+            for(auto &it : proto->vars){
+                if(ctx == NULL){
+                    items.push_back(it.second);
+                }else{
+                    ctx->set(it.first, it.second);
+                }
+            }
+        }  
+    }else
+    if(effects.expand && solved->type != CV::ItemTypes::CONTEXT && solved->type != CV::ItemTypes::LIST){
+        cursor->setError("'"+CV::ModifierTypes::str(CV::ModifierTypes::EXPAND)+"': Expand modifier can only be used on iterables.");
+        items.push_back(solved);
+    }else{   
+        items.push_back(solved);
+    } 
+}
+
+static std::shared_ptr<CV::Item> processIteration(const std::vector<CV::Token> &tokens, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+    std::vector<std::shared_ptr<CV::Item>> items;
+    for(int i = 0; i < tokens.size(); ++i){
+        CV::ModifierEffect effects;
+        auto interp = interpret(tokens[i], cursor, ctx);
+        auto solved = processPreInterpretModifiers(interp, tokens[i].modifiers, cursor, ctx, effects);
+        if(cursor->error){
+            return std::make_shared<CV::Item>(CV::Item());
+        }
+        if(solved->type == CV::ItemTypes::INTERRUPT){
+            auto interrupt =  std::static_pointer_cast<CV::Interrupt>(solved);
+            if(interrupt->intType == CV::InterruptTypes::SKIP){
+                if(interrupt->hasPayload()){
+                    items.push_back(interrupt->getPayload());
+                }
+                continue;
+            }else
+            if(interrupt->intType == CV::InterruptTypes::STOP){
+                return std::static_pointer_cast<CV::Item>(interrupt);
+            }
+        }   
+        processPostInterpretExpandListOrContext(solved, effects, items, cursor, NULL);
+        if(cursor->error){
+            return std::make_shared<CV::Item>(CV::Item());
+        }                           
+    }          
+    return tokens.size() == 1 ? items[0] : std::make_shared<CV::List>(CV::List(items));    
+}
+
+static bool checkCond(std::shared_ptr<CV::Item> &item){
+    if(item->type == CV::ItemTypes::NUMBER){
+        return std::static_pointer_cast<CV::Number>(item)->n <= 0 ? false : true;
+    }else
+    if(item->type != CV::ItemTypes::NIL){
+        return true;
+    }else{
+        return false;
+    }
+};
+
+static std::shared_ptr<CV::Item> eval(const std::vector<CV::Token> &tokens, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+
+    auto first = tokens[0];
+    auto imp = first.first;
+
+    if(imp == "set"){
+        auto params = compileTokens(tokens, 1, tokens.size()-1);
+        if(params.size() > 2 || params.size() < 2){
+            cursor->setError("'set': Expects exactly 2 arguments: <NAME> <VALUE>.");
+            return std::make_shared<CV::Item>(CV::Item());
+        }
+        CV::ModifierEffect effects;
+        auto interp = interpret(params[1], cursor, ctx);
+        auto solved = processPreInterpretModifiers(interp, params[1].modifiers, cursor, ctx, effects);
+        if(cursor->error){
+            return std::make_shared<CV::Item>(CV::Item());
+        }
+        auto f = ctx->getWithContext(params[0].first);
+        if(f.item){
+            return f.context->set(params[0].first, solved);
+        }else{
+
+            return ctx->set(params[0].first, solved);
+        }
+    }else
+    if(imp == "skip" || imp == "stop"){
+        auto interrupt = std::make_shared<CV::Interrupt>(CV::Interrupt(imp == "skip" ? CV::InterruptTypes::SKIP : CV::InterruptTypes::STOP));
+        if(tokens.size() > 1){
+            auto params = compileTokens(tokens, 1, tokens.size()-1);
+            auto solved = processIteration(params, cursor, ctx);
+            if(solved->type == CV::ItemTypes::INTERRUPT){
+                return solved;
+            }else{
+                interrupt->setPayload(solved);
+            }
+        }
+        return std::static_pointer_cast<CV::Item>(interrupt);
+    }else
+    if(imp == "if"){
+        if(tokens.size() > 4){
+            cursor->setError("'if': Expects no more than 3 arguments: <CONDITION> <[CODE[0]...CODE[n]]> (ELSE-OPT)<[CODE[0]...CODE[n]]>.");
+            return std::make_shared<CV::Item>(CV::Item());
+        }
+        if(tokens.size() < 3){
+            cursor->setError("'if': Expects no less than 2 arguments: <CONDITION> <[CODE[0]...CODE[n]]> (ELSE-OPT)<[CODE[0]...CODE[n]]>.");
+            return std::make_shared<CV::Item>(CV::Item());
+        }        
+
+        auto params = compileTokens(tokens, 1, tokens.size()-1);
+        auto cond = params[0];
+        auto trueBranch = params[1];
+        auto tctx = std::make_shared<CV::Context>(CV::Context(ctx));
+        tctx->setTop(ctx);
+
+        CV::ModifierEffect origEff;
+        auto interp = interpret(cond, cursor, tctx);
+        auto condV = processPreInterpretModifiers(interp, cond.modifiers, cursor, tctx, origEff);
+        if(cursor->error){
+            return std::make_shared<CV::Item>(CV::Item());
+        } 
+        if(checkCond(condV)){
+            CV::ModifierEffect effects;
+            auto interp = interpret(trueBranch, cursor, tctx);
+            auto solved = processPreInterpretModifiers(interp, trueBranch.modifiers, cursor, tctx, effects);
+            if(cursor->error){
+                return std::make_shared<CV::Item>(CV::Item());
+            }
+            return solved;      
+        }else
+        if(params.size() == 3){
+            auto falseBranch = params[2];
+            CV::ModifierEffect effects;
+            auto interp = interpret(falseBranch, cursor, tctx);
+            auto solved = processPreInterpretModifiers(interp, falseBranch.modifiers, cursor, tctx, effects);
+            if(cursor->error){
+                return std::make_shared<CV::Item>(CV::Item());
+            }
+            return solved;            
+        }else{
+            return condV;
+        }
+    }else
+    if(imp == "do"){
+        if(tokens.size() > 3 || tokens.size() < 3){
+            cursor->setError("'do': Expects exactly 2 arguments: <CONDITION> <[CODE[0]...CODE[n]]>.");
+            return std::make_shared<CV::Item>(CV::Item());
+        }
+        auto params = compileTokens(tokens, 1, tokens.size()-1);
+        auto cond = params[0];
+        auto code = params[1];
+        auto tctx = std::make_shared<CV::Context>(CV::Context(ctx));
+        tctx->setTop(ctx);
+
+        CV::ModifierEffect origEff;
+        auto interp = interpret(cond, cursor, tctx);
+        auto condV = processPreInterpretModifiers(interp, cond.modifiers, cursor, tctx, origEff);
+        if(cursor->error){
+            return std::make_shared<CV::Item>(CV::Item());
+        }  
+        auto last = std::shared_ptr<CV::Item>(NULL);
+        while(checkCond(condV)){
+            CV::ModifierEffect effects;
+            auto interp = interpret(code, cursor, tctx);
+            auto solved = processPreInterpretModifiers(interp, code.modifiers, cursor, tctx, effects);
+            if(cursor->error){
+                return std::make_shared<CV::Item>(CV::Item());
+            }
+            if(solved->type == CV::ItemTypes::INTERRUPT){
+                auto interrupt =  std::static_pointer_cast<CV::Interrupt>(solved);
+                if(interrupt->intType == CV::InterruptTypes::SKIP){
+                    continue;
+                }else
+                if(interrupt->intType == CV::InterruptTypes::STOP){
+                    return interrupt->hasPayload() ?  interrupt->getPayload() : interrupt;
+                }
+            }         
+            last = solved;
+            // Next iteration
+            auto reinterp = interpret(cond, cursor, tctx);
+            condV = processPreInterpretModifiers(reinterp, cond.modifiers, cursor, tctx, origEff);
+            if(cursor->error){
+                return std::make_shared<CV::Item>(CV::Item());
+            }            
+        }             
+        return last ? last : std::make_shared<CV::Item>(CV::Item());
+    }else
+    if(imp == "iter"){
+        if(tokens.size() > 3 || tokens.size() < 3){
+            cursor->setError("'iter': Expects exactly 2 arguments: <ITERABLE> <[CODE[0]...CODE[n]]>.");
+            return std::make_shared<CV::Item>(CV::Item());
+        }        
+        auto params = compileTokens(tokens, 1, tokens.size()-1);
+        auto code = params[1];
+        auto tctx = std::make_shared<CV::Context>(CV::Context(ctx));
+        tctx->setTop(ctx);
+        CV::ModifierEffect origEff;
+        auto interp = interpret(params[0], cursor, tctx);
+        auto iterable = processPreInterpretModifiers(interp, params[0].modifiers, cursor, tctx, origEff);
+        if(cursor->error){
+            return std::make_shared<CV::Item>(CV::Item());
+        }            
+        if(iterable->type == CV::ItemTypes::LIST){
+            std::vector<std::shared_ptr<CV::Item>> items;
+            auto list = std::static_pointer_cast<CV::List>(iterable);
+            for(int i = 0; i < list->data.size(); ++i){
+                auto item = list->get(i);
+                // We use the namer to name every iteration's current variable
+                if(origEff.named.size() > 0){
+                    tctx->set(origEff.named, item);
+                }                
+                CV::ModifierEffect effects;
+                auto interp = interpret(code, cursor, tctx);
+                auto solved = processPreInterpretModifiers(interp, code.modifiers, cursor, tctx, effects);
+                if(cursor->error){
+                    return std::make_shared<CV::Item>(CV::Item());
+                }                   
+                if(solved->type == CV::ItemTypes::INTERRUPT){
+                    auto interrupt =  std::static_pointer_cast<CV::Interrupt>(solved);
+                    if(interrupt->intType == CV::InterruptTypes::SKIP){
+                        if(interrupt->hasPayload()){
+                            items.push_back(interrupt->getPayload());
+                        }
+                        continue;
+                    }else
+                    if(interrupt->intType == CV::InterruptTypes::STOP){
+                        return interrupt->hasPayload() ?  interrupt->getPayload() : interrupt;
+                    }
+                }           
+                processPostInterpretExpandListOrContext(solved, effects, items, cursor, NULL);
+                if(cursor->error){
+                    return std::make_shared<CV::Item>(CV::Item());
+                }                               
+            }
+            return std::static_pointer_cast<CV::Item>(std::make_shared<CV::List>(CV::List(items)));
+        }else
+        if(iterable->type == CV::ItemTypes::CONTEXT){
+            std::vector<std::shared_ptr<CV::Item>> items;
+            auto list = std::static_pointer_cast<CV::Context>(iterable);
+            for(auto &it : list->vars){
+                auto item = it.second;
+                // We use the namer to name every iteration's current variable
+                if(origEff.named.size() > 0){
+                    tctx->set(origEff.named, item);
+                }                
+                CV::ModifierEffect effects;
+                auto interp = interpret(code, cursor, tctx);
+                auto solved = processPreInterpretModifiers(interp, code.modifiers, cursor, tctx, effects);
+                if(cursor->error){
+                    return std::make_shared<CV::Item>(CV::Item());
+                }             
+                if(solved->type == CV::ItemTypes::INTERRUPT){
+                    auto interrupt =  std::static_pointer_cast<CV::Interrupt>(solved);
+                    if(interrupt->intType == CV::InterruptTypes::SKIP){
+                        if(interrupt->hasPayload()){
+                            items.push_back(interrupt->getPayload());
+                        }
+                        continue;
+                    }else
+                    if(interrupt->intType == CV::InterruptTypes::STOP){
+                        return interrupt->hasPayload() ?  interrupt->getPayload() : interrupt;
+                    }
+                }               
+                processPostInterpretExpandListOrContext(solved, effects, items, cursor, NULL);
+                if(cursor->error){
+                    return std::make_shared<CV::Item>(CV::Item());
+                } 
+            }
+            return std::static_pointer_cast<CV::Item>(std::make_shared<CV::List>(CV::List(items)));
+        }else{
+            cursor->setError("'iter': Expects argument(0) to be iterable.");
+            return std::make_shared<CV::Item>(CV::Item());        
+        }
+    }else
+    if(imp == "fn"){
+        auto params = compileTokens(tokens, 1, tokens.size()-1);
+        if(params.size() > 2 || params.size() < 2){
+            cursor->setError("'fn': Expects exactly 2 arguments: <[ARGS[0]...ARGS[n]]> <[CODE[0]...CODE[n]]>.");
+            return std::make_shared<CV::Item>(CV::Item());
+        }
+        bool isVariadic = false;
+        auto argNames = parseTokens(params[0].first, ' ', cursor);
+        if(cursor->error){
+            return std::make_shared<CV::Item>(CV::Item());
+        }
+        if(argNames.size() == 1 && argNames[0] == "..."){
+            isVariadic = true;
+        }
+        auto code = params[1];
+
+        return std::static_pointer_cast<CV::Item>(  std::make_shared<CV::Function>(isVariadic ? std::vector<std::string>({}) : argNames, code.first, isVariadic)  );
+    }else
+    if(imp == "ct"){
+        auto params = compileTokens(tokens, 1, tokens.size()-1);
+        auto pctx = std::make_shared<CV::Context>(CV::Context(ctx));
+        for(int i = 0; i < params.size(); ++i){
+            CV::ModifierEffect effects;
+            auto interp = interpret(params[i], cursor, pctx);
+            auto solved = processPreInterpretModifiers(interp, params[i].modifiers, cursor, pctx, effects);
+            if(cursor->error){
+                return std::make_shared<CV::Item>(CV::Item());
+            }                
+            std::vector<std::shared_ptr<CV::Item>> items;
+            processPostInterpretExpandListOrContext(solved, effects, items, cursor, pctx);
+            if(cursor->error){
+                return std::make_shared<CV::Item>(CV::Item());
+            }  
+            for(int j = 0; j < items.size(); ++j){
+                if(effects.named.size() == 0){
+                    pctx->set("v-"+std::to_string(j), items[j]);
+                }
+            }
+        }
+        return pctx;
+    }else{
+        auto var = ctx->get(imp);
+        // Is it a function?
+        if(var && var->type == CV::ItemTypes::FUNCTION){
+            auto fn = std::static_pointer_cast<CV::Function>(var);
+            return runFunction(fn, compileTokens(tokens, 1, tokens.size()-1), cursor, ctx);
+        }else   
+        // Is it just a variable?
+        if(var && tokens.size() == 1){
+            CV::ModifierEffect effects;
+            auto r = processPreInterpretModifiers(var, first.modifiers, cursor, ctx, effects);
+            return r;
+        }else
+        // Is it a natural type?
+        if(tokens.size() == 1){
+            if(!var && CV::Tools::isNumber(imp)){
+                auto n = std::make_shared<CV::Number>(CV::Number(std::stod(imp)));
+                CV::ModifierEffect effects;
+                auto casted = std::static_pointer_cast<CV::Item>(n);
+                auto result = processPreInterpretModifiers(casted, first.modifiers, cursor, ctx, effects);
+                if(cursor->error){
+                    return std::make_shared<CV::Item>(CV::Item());
+                }                      
+                return result;
+            }else
+            if(!var && CV::Tools::isString(imp)){
+                CV::ModifierEffect effects;
+                auto n = std::make_shared<CV::String>(CV::String(imp.substr(1, imp.length() - 2)));
+                auto casted = std::static_pointer_cast<CV::Item>(n);
+                auto result = processPreInterpretModifiers(casted, first.modifiers, cursor, ctx, effects);
+                return result;
+            }else
+            if(!var && CV::Tools::isList(imp)){
+                CV::ModifierEffect effects;
+                auto interp = interpret(first, cursor, ctx);
+                auto result = processPreInterpretModifiers(interp, first.modifiers,cursor, ctx, effects);
+                if(cursor->error){
+                    return std::make_shared<CV::Item>(CV::Item());
+                }                   
+                return result;
+            }else
+            if(imp == "nil"){
+                CV::ModifierEffect effects;
+                auto n = std::make_shared<CV::Item>(CV::Item());
+                auto result = processPreInterpretModifiers(n, first.modifiers, cursor, ctx, effects);
+                if(cursor->error){
+                    return std::make_shared<CV::Item>(CV::Item());
+                }                      
+                return result;
+            }else{
+                cursor->setError("'"+imp+"': Undefined symbol or imperative."); // TODO: Check if it's running as relaxed
+                return std::make_shared<CV::Item>(CV::Item());
+            }
+        }else
+        // List or in-line context switch?
+        if(tokens.size() > 1){
+            auto fmods = tokens[0].getModifier(CV::ModifierTypes::SCOPE);
+            // Super rad inline-contex switching d:)
+            if(fmods.type == CV::ModifierTypes::SCOPE){
+                CV::ModifierEffect effects;
+                auto interp = interpret(tokens[0], cursor, ctx);
+                auto solved = processPreInterpretModifiers(interp, tokens[0].modifiers, cursor, ctx, effects);
+                if(cursor->error){
+                    return std::make_shared<CV::Item>(CV::Item());
+                } 
+                auto inlineCode = compileTokens(tokens, 1, tokens.size()-1);
+                return eval(inlineCode, cursor, effects.toCtx); // Check if toCtx is null? Shouldn't be possible tho
+            }else{
+            // List it is
+                return processIteration(tokens, cursor, ctx);
+            }
+        }
+    }
+
+    return std::make_shared<CV::Item>(CV::Item());
+}
+
+static bool typicalVariadicArithmeticCheck(const std::string &name, const std::vector<std::shared_ptr<CV::Item>> &params, CV::Cursor *cursor, int max = -1){
+    CV::FunctionConstraints consts;
+    consts.setMinParams(2);
+    if(max != -1){
+        consts.setMaxParams(max);
+    }
+    consts.setExpectType(CV::ItemTypes::NUMBER);
+    consts.allowMisMatchingTypes = false;
+    consts.allowNil = false;
+    std::string errormsg;
+    if(!consts.test(params, errormsg)){
+        cursor->setError("'"+name+"': "+errormsg);
+        return false;
+    }
+    return true;
+};
+
+void CV::AddStandardOperators(std::shared_ptr<CV::Context> &ctx){
+
+    /*
+
+        ARITHMETIC OPERATORS
+
+    */
+    ctx->set("+", std::make_shared<CV::Function>(CV::Function({}, [](const std::vector<std::shared_ptr<CV::Item>> &params, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+        if(!typicalVariadicArithmeticCheck("+", params, cursor)){
+            return std::make_shared<CV::Item>(CV::Item());
+        }
+        __CV_NUMBER_NATIVE_TYPE r = std::static_pointer_cast<CV::Number>(params[0])->n;
+        for(int i = 1; i < params.size(); ++i){
+            r += std::static_pointer_cast<CV::Number>(params[i])->n;
+        }
+        return std::static_pointer_cast<CV::Item>(std::make_shared<CV::Number>(CV::Number(r)));
+    }, true)));
+    ctx->set("-", std::make_shared<CV::Function>(CV::Function({}, [](const std::vector<std::shared_ptr<CV::Item>> &params, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+        if(!typicalVariadicArithmeticCheck("-", params, cursor)){
+            return std::make_shared<CV::Item>(CV::Item());
+        }
+        __CV_NUMBER_NATIVE_TYPE r = std::static_pointer_cast<CV::Number>(params[0])->n;
+        for(int i = 1; i < params.size(); ++i){
+            r -= std::static_pointer_cast<CV::Number>(params[i])->n;
+        }
+        return std::static_pointer_cast<CV::Item>(std::make_shared<CV::Number>(CV::Number(r)));
+    }, true)));
+    ctx->set("*", std::make_shared<CV::Function>(CV::Function({}, [](const std::vector<std::shared_ptr<CV::Item>> &params, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+        if(!typicalVariadicArithmeticCheck("*", params, cursor)){
+            return std::make_shared<CV::Item>(CV::Item());
+        }
+        __CV_NUMBER_NATIVE_TYPE r = std::static_pointer_cast<CV::Number>(params[0])->n;
+        for(int i = 1; i < params.size(); ++i){
+            r *= std::static_pointer_cast<CV::Number>(params[i])->n;
+        }
+        return std::static_pointer_cast<CV::Item>(std::make_shared<CV::Number>(CV::Number(r)));
+    }, true)));  
+    ctx->set("/", std::make_shared<CV::Function>(CV::Function({}, [](const std::vector<std::shared_ptr<CV::Item>> &params, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+        if(!typicalVariadicArithmeticCheck("/", params, cursor)){
+            return std::make_shared<CV::Item>(CV::Item());
+        }
+        __CV_NUMBER_NATIVE_TYPE r = std::static_pointer_cast<CV::Number>(params[0])->n;
+        for(int i = 1; i < params.size(); ++i){
+            r /= std::static_pointer_cast<CV::Number>(params[i])->n;
+        }
+        return std::static_pointer_cast<CV::Item>(std::make_shared<CV::Number>(CV::Number(r)));
+    }, true)));
+    ctx->set("pow", std::make_shared<CV::Function>(Function({"a", "b"}, [](const std::vector<std::shared_ptr<CV::Item>> &params, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+        if(!typicalVariadicArithmeticCheck("pow", params, cursor, 2)){
+            return std::make_shared<CV::Item>(CV::Item());
+        }
+        __CV_NUMBER_NATIVE_TYPE r = std::static_pointer_cast<CV::Number>(params[0])->n;
+        for(int i = 1; i < params.size(); ++i){
+            r =  std::pow(r, std::static_pointer_cast<CV::Number>(params[i])->n);
+        }
+        return std::static_pointer_cast<CV::Item>(std::make_shared<CV::Number>(CV::Number(r)));
+    }, false))); 
+    ctx->set("sqrt", std::make_shared<CV::Function>(CV::Function({}, [](const std::vector<std::shared_ptr<CV::Item>> &params, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+        CV::FunctionConstraints consts;
+        consts.setMinParams(1);
+        consts.setExpectType(CV::ItemTypes::NUMBER);
+        consts.allowMisMatchingTypes = false;
+        consts.allowNil = false;
+        std::string errormsg;
+        if(!consts.test(params, errormsg)){
+            cursor->setError("'sqrt': "+errormsg);
+            return std::make_shared<CV::Item>(CV::Item());
+        }
+        if(params.size() == 1){
+            return std::static_pointer_cast<CV::Item>(std::make_shared<CV::Number>(CV::Number(  std::sqrt(std::static_pointer_cast<CV::Number>(params[0])->n)   )));
+        }else{
+            std::vector<std::shared_ptr<CV::Item>> result;
+            for(int i = 0; i < params.size(); ++i){
+                result.push_back(
+                    std::static_pointer_cast<CV::Item>(std::make_shared<CV::Number>(CV::Number(  std::sqrt(std::static_pointer_cast<CV::Number>(params[i])->n)   )))
+                );
+            }
+            return std::static_pointer_cast<CV::Item>(std::make_shared<CV::List>(CV::List(result, false)));
+        }
+    }, true)));    
+       
+    /* 
+        QUICK MATH
+    */ 
+    ctx->set("++", std::make_shared<CV::Function>(CV::Function({}, [](const std::vector<std::shared_ptr<CV::Item>> &params, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+        CV::FunctionConstraints consts;
+        consts.setMinParams(1);
+        consts.allowMisMatchingTypes = false;        
+        consts.allowNil = false;
+        consts.setExpectType(CV::ItemTypes::NUMBER);
+
+        std::string errormsg;
+        if(!consts.test(params, errormsg)){
+            cursor->setError("'++': "+errormsg);
+            return std::make_shared<CV::Item>(CV::Item());
+        }
+
+        if(params.size() == 1){
+            ++std::static_pointer_cast<CV::Number>(params[0])->n;
+            return params[0];
+        }else{   
+            for(int i = 0; i < params.size(); ++i){
+                ++std::static_pointer_cast<CV::Number>(params[i])->n;
+            }
+        }
+
+        return std::static_pointer_cast<CV::Item>(std::make_shared<CV::List>(CV::List(params, false)));
+    }, true)));
+
+    ctx->set("--", std::make_shared<CV::Function>(CV::Function({}, [](const std::vector<std::shared_ptr<CV::Item>> &params, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+        CV::FunctionConstraints consts;
+        consts.setMinParams(1);
+        consts.allowMisMatchingTypes = false;        
+        consts.allowNil = false;
+        consts.setExpectType(CV::ItemTypes::NUMBER);
+
+        std::string errormsg;
+        if(!consts.test(params, errormsg)){
+            cursor->setError("'--': "+errormsg);
+            return std::make_shared<CV::Item>(CV::Item());
+        }
+
+        if(params.size() == 1){
+            --std::static_pointer_cast<CV::Number>(params[0])->n;
+            return params[0];
+        }else{   
+            for(int i = 0; i < params.size(); ++i){
+                --std::static_pointer_cast<CV::Number>(params[i])->n;
+            }
+        }
+
+        return std::static_pointer_cast<CV::Item>(std::make_shared<CV::List>(CV::List(params, false)));
+    }, true)));    
+
+
+    ctx->set("//", std::make_shared<CV::Function>(CV::Function({}, [](const std::vector<std::shared_ptr<CV::Item>> &params, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+        CV::FunctionConstraints consts;
+        consts.setMinParams(1);
+        consts.allowMisMatchingTypes = false;        
+        consts.allowNil = false;
+        consts.setExpectType(CV::ItemTypes::NUMBER);
+
+        std::string errormsg;
+        if(!consts.test(params, errormsg)){
+            cursor->setError("'//': "+errormsg);
+            return std::make_shared<CV::Item>(CV::Item());
+        }
+
+        if(params.size() == 1){
+            std::static_pointer_cast<CV::Number>(params[0])->n /= 2;
+            return params[0];
+        }else{   
+            for(int i = 0; i < params.size(); ++i){
+                std::static_pointer_cast<CV::Number>(params[i])->n /= 2;
+            }
+        }
+
+        return std::static_pointer_cast<CV::Item>(std::make_shared<CV::List>(CV::List(params, false)));
+    }, true)));    
+
+    ctx->set("**", std::make_shared<CV::Function>(CV::Function({}, [](const std::vector<std::shared_ptr<CV::Item>> &params, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+        CV::FunctionConstraints consts;
+        consts.setMinParams(1);
+        consts.allowMisMatchingTypes = false;        
+        consts.allowNil = false;
+        consts.setExpectType(CV::ItemTypes::NUMBER);
+
+        std::string errormsg;
+        if(!consts.test(params, errormsg)){
+            cursor->setError("'**': "+errormsg);
+            return std::make_shared<CV::Item>(CV::Item());
+        }
+
+        if(params.size() == 1){
+            std::static_pointer_cast<CV::Number>(params[0])->n *= std::static_pointer_cast<CV::Number>(params[0])->n;
+            return params[0];
+        }else{   
+            for(int i = 0; i < params.size(); ++i){
+                std::static_pointer_cast<CV::Number>(params[i])->n *= std::static_pointer_cast<CV::Number>(params[i])->n;
+            }
+        }
+
+        return std::static_pointer_cast<CV::Item>(std::make_shared<CV::List>(CV::List(params, false)));
+    }, true)));
+
+
+    /*
+        LISTS
+    */
+    ctx->set("l-range", std::make_shared<CV::Function>(CV::Function({}, [](const std::vector<std::shared_ptr<CV::Item>> &params, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+        if(!typicalVariadicArithmeticCheck("l-range", params, cursor)){
+            return std::make_shared<CV::Item>(CV::Item());
+        }
+
+        auto s = std::static_pointer_cast<CV::Number>(params[0])->n;
+        auto e = std::static_pointer_cast<CV::Number>(params[1])->n;
+        
+        std::vector<std::shared_ptr<CV::Item>> result;
+        if(s == e){
+            return std::static_pointer_cast<CV::Item>(std::make_shared<CV::List>(CV::List(result, false)));
+        }
+
+        bool inc = e > s;
+        __CV_NUMBER_NATIVE_TYPE step = Tools::positive(params.size() == 3 ? std::static_pointer_cast<CV::Number>(params[2])->n : 1);
+
+        for(__CV_NUMBER_NATIVE_TYPE i = s; (inc ? i <= e : i >= e); i += (inc ? step : -step)){
+            result.push_back(
+                std::static_pointer_cast<CV::Item>(std::make_shared<CV::Number>(i))
+            );
+        }
+
+        return std::static_pointer_cast<CV::Item>(std::make_shared<CV::List>(CV::List(result, false)));
+    }, true)));
+    
+    ctx->set("l-sort", std::make_shared<CV::Function>(Function({"c", "l"}, [](const std::vector<std::shared_ptr<CV::Item>> &params, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+        CV::FunctionConstraints consts;
+        consts.setMinParams(2);
+        consts.setMaxParams(2);
+        consts.allowMisMatchingTypes = true;        
+        consts.allowNil = false;
+        consts.setExpectedTypeAt(0, CV::ItemTypes::STRING);
+        consts.setExpectedTypeAt(1, CV::ItemTypes::LIST);
+        
+        // TODO: add function support
+
+        std::string errormsg;
+        if(!consts.test(params, errormsg)){
+            cursor->setError("'l-sort': "+errormsg);
+            return std::make_shared<CV::Item>(CV::Item());
+        }
+
+        auto criteria = std::static_pointer_cast<CV::String>(params[0]);
+
+        auto &order = criteria->data;
+        if(order != "DESC" && order != "ASC"){
+            cursor->setError("'l-sort': criteria can only be 'DESC' or 'ASC'.");
+            return std::make_shared<CV::Item>(CV::Item());
+        }        
+
+        auto list = std::static_pointer_cast<CV::List>(params[1]);
+        std::vector<std::shared_ptr<CV::Item>> items;
+        for(int i = 0; i < list->data.size(); ++i){
+            items.push_back(list->get(i));
+        }        
+
+        CV::FunctionConstraints constsList;
+        constsList.setExpectType(CV::ItemTypes::NUMBER);
+        constsList.allowMisMatchingTypes = false;        
+        constsList.allowNil = false;
+        if(!constsList.test(items, errormsg)){
+            cursor->setError("'l-sort': "+errormsg);
+            return std::make_shared<CV::Item>(CV::Item());
+        }        
+
+        std::sort(items.begin(), items.end(), [&](std::shared_ptr<CV::Item> &a, std::shared_ptr<CV::Item> &b){
+            return order == "DESC" ?     std::static_pointer_cast<CV::Number>(a)->n > std::static_pointer_cast<CV::Number>(b)->n :
+                                        std::static_pointer_cast<CV::Number>(a)->n < std::static_pointer_cast<CV::Number>(b)->n;
+        });
+
+        return std::static_pointer_cast<CV::Item>(std::make_shared<CV::List>(CV::List(items, false)));
+    }, false)));    
+
+
+    ctx->set("l-filter", std::make_shared<CV::Function>(Function({"c", "l"}, [](const std::vector<std::shared_ptr<CV::Item>> &params, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+        CV::FunctionConstraints consts;
+        consts.setMinParams(2);
+        consts.setMaxParams(2);
+        consts.allowMisMatchingTypes = true;        
+        consts.allowNil = false;
+        consts.setExpectedTypeAt(0, CV::ItemTypes::FUNCTION);
+        consts.setExpectedTypeAt(1, CV::ItemTypes::LIST);
+        
+        std::string errormsg;
+        if(!consts.test(params, errormsg)){
+            cursor->setError("'l-filter': "+errormsg);
+            return std::make_shared<CV::Item>(CV::Item());
+        }
+
+        auto criteria = std::static_pointer_cast<Function>(params[0]);
+
+        if(criteria->variadic || criteria->params.size() != 1){
+            cursor->setError("'l-filter': criteria's function must receive exactly 1 argument(a).");
+            return std::make_shared<CV::Item>(CV::Item());
+        }
+
+        auto list = std::static_pointer_cast<CV::List>(params[1]);
+        std::vector<std::shared_ptr<CV::Item>> items;
+        for(int i = 0; i < list->data.size(); ++i){
+            auto result = runFunction(criteria, std::vector<std::shared_ptr<CV::Item>>{list->get(i)}, cursor, ctx);
+            if(checkCond(result)){
+                continue;
+            }
+            items.push_back(list->get(i));
+        }        
+
+        return std::static_pointer_cast<CV::Item>(std::make_shared<CV::List>(CV::List(items, false)));
+
+    }, false)));  
+
+
+    ctx->set("l-flat", std::make_shared<CV::Function>(CV::Function({}, [](const std::vector<std::shared_ptr<CV::Item>> &params, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+        CV::FunctionConstraints consts;
+        consts.allowMisMatchingTypes = true;        
+        consts.allowNil = true;
+        
+        std::string errormsg;
+        if(!consts.test(params, errormsg)){
+            cursor->setError("'l-flat': "+errormsg);
+            return std::make_shared<CV::Item>(CV::Item());
+        }
+
+
+        std::vector<std::shared_ptr<CV::Item>> finalList;
+
+        std::function<void(std::shared_ptr<CV::Item> &item)> recursive = [&](const std::shared_ptr<CV::Item> &item){
+            if(item->type == CV::ItemTypes::LIST){
+                auto list = std::static_pointer_cast<CV::List>(item);
+                for(int i = 0; i < list->data.size(); ++i){
+                    auto item = list->get(i);
+                    recursive(item);
+                }
+            }else
+            if(item->type == CV::ItemTypes::CONTEXT){
+                auto ctx = std::static_pointer_cast<CV::Context>(item);
+                for(auto &it : ctx->vars){
+                    recursive(it.second);
+                }
+            }else{                
+                finalList.push_back(item);
+            }
+        };
+
+
+        for(int i = 0; i < params.size(); ++i){
+            auto item = params[i];
+            recursive(item);
+        }
+
+
+
+        return std::static_pointer_cast<CV::Item>(std::make_shared<CV::List>(CV::List(finalList, false)));
+    }, true)));   
+
+
+
+    ctx->set("splice", std::make_shared<CV::Function>(CV::Function({}, [](const std::vector<std::shared_ptr<CV::Item>> &params, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+        CV::FunctionConstraints consts;
+        consts.setMinParams(2);
+        consts.allowNil = false;
+        consts.allowMisMatchingTypes = false;
+        consts.setExpectType(CV::ItemTypes::CONTEXT);
+        consts.setExpectType(CV::ItemTypes::LIST);
+
+        std::string errormsg;
+        if(!consts.test(params, errormsg)){
+            cursor->setError("'splice': "+errormsg);
+            return std::make_shared<CV::Item>(CV::Item());
+        }
+
+        auto getAllItems = [](std::shared_ptr<CV::Item> &item){
+            std::unordered_map<std::string, std::shared_ptr<CV::Item>> all;
+            if(item->type == CV::ItemTypes::LIST){
+                auto list = std::static_pointer_cast<CV::List>(item);
+                for(int i = 0; i < list->data.size(); ++i){
+                    all[std::to_string(i)] = list->get(i);
+                }
+            }else{
+                auto ctx = std::static_pointer_cast<CV::Context>(item);
+                for(auto &it : ctx->vars){
+                    all[it.first] = it.second;
+                }
+            }
+            return all;
+        };
+
+        auto first = params[0];
+        // First item decides the spliced result
+        if(first->type == CV::ItemTypes::LIST){
+            std::vector<std::shared_ptr<CV::Item>> compiled;
+            for(int i = 0; i < params.size(); ++i){
+                auto current = params[i];
+                auto all = getAllItems(current);
+                for(auto &it : all){
+                    compiled.push_back(it.second);
+                }
+            }
+            return std::static_pointer_cast<CV::Item>(std::make_shared<CV::List>(CV::List(compiled, false)));
+        }else{
+            auto ctx = std::make_shared<CV::Context>(CV::Context());
+            for(int i = 0; i < params.size(); ++i){
+                auto current = params[i];
+                auto all = getAllItems(current);
+                for(auto &it : all){
+                    ctx->set(it.first, it.second);
+                }
+            }
+            return std::static_pointer_cast<CV::Item>(ctx);
+        }
+
+        return std::make_shared<CV::Item>(CV::Item());
+    }, true)));    
+
+    ctx->set("l-push", std::make_shared<CV::Function>(CV::Function({}, [](const std::vector<std::shared_ptr<CV::Item>> &params, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+        CV::FunctionConstraints consts;
+        consts.setMinParams(2);
+        consts.allowMisMatchingTypes = true;
+        consts.allowNil = true;
+
+        std::string errormsg;
+        if(!consts.test(params, errormsg)){
+            cursor->setError("'l-push': "+errormsg);
+            return std::make_shared<CV::Item>(CV::Item());
+        }
+
+        std::vector<std::shared_ptr<CV::Item>> items;
+        auto last = params[params.size()-1];
+
+        if(last->type == CV::ItemTypes::LIST){
+            auto list = std::static_pointer_cast<CV::List>(last);
+            for(int i = 0; i < list->data.size(); ++i){
+                items.push_back(list->get(i));
+            }
+        }else{
+            items.push_back(last);
+        }
+
+        for(int i = 0; i < params.size()-1; ++i){
+            items.push_back(params[i]);
+        }
+
+        return std::static_pointer_cast<CV::Item>(std::make_shared<CV::List>(CV::List(items, false)));
+
+    }, true))); 
+    
+    ctx->set("l-pop", std::make_shared<CV::Function>(CV::Function({}, [](const std::vector<std::shared_ptr<CV::Item>> &params, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+        CV::FunctionConstraints consts;
+        consts.setMinParams(2);
+        consts.allowMisMatchingTypes = true;
+        consts.allowNil = true;
+
+        std::string errormsg;
+        if(!consts.test(params, errormsg)){
+            cursor->setError("'<<': "+errormsg);
+            return std::make_shared<CV::Item>(CV::Item());
+        }
+
+        std::vector<std::shared_ptr<CV::Item>> items;
+        auto first = params[0];
+
+        if(first->type == CV::ItemTypes::LIST){
+            auto list = std::static_pointer_cast<CV::List>(first);
+            for(int i = 0; i < list->data.size(); ++i){
+                items.push_back(list->get(i));
+            }
+        }else{
+            items.push_back(first);
+        }
+
+        for(int i = 1; i < params.size(); ++i){
+            items.push_back(params[i]);
+        }
+
+        return std::static_pointer_cast<CV::Item>(std::make_shared<CV::List>(CV::List(items, false)));
+
+    }, true)));     
+
+    ctx->set("l-sub", std::make_shared<CV::Function>(Function({"l", "p", "n"}, [](const std::vector<std::shared_ptr<CV::Item>> &params, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+        CV::FunctionConstraints consts;
+        consts.setMinParams(2);
+        consts.setMaxParams(3);
+        consts.allowMisMatchingTypes = true;
+        consts.allowNil = false;
+        consts.setExpectedTypeAt(0, CV::ItemTypes::LIST);
+        consts.setExpectedTypeAt(1, CV::ItemTypes::NUMBER);
+        consts.setExpectedTypeAt(2, CV::ItemTypes::NUMBER);
+        static const std::string name = "l-sub";
+        std::string errormsg;
+        if(!consts.test(params, errormsg)){
+            cursor->setError("'"+name+"': "+errormsg);
+            return std::make_shared<CV::Item>(CV::Item());
+        }
+
+        int index = std::static_pointer_cast<CV::Number>(params[1])->n;
+        auto list = std::static_pointer_cast<CV::List>(params[0]);
+        int n = params.size() > 2 ? std::static_pointer_cast<CV::Number>(params[2])->n : 1;
+
+        if(index+n < 0 || index+n > list->data.size()){
+            cursor->setError("'"+name+"': position("+std::to_string(index+n)+") is out of range("+std::to_string(list->data.size())+")");
+            return std::make_shared<CV::Item>(CV::Item());
+        }      
+
+        std::vector<std::shared_ptr<CV::Item>> result;
+        for(int i = 0; i < list->data.size(); ++i){
+            if(i >= index && i < index+n){
+                result.push_back(list->get(i));
+            }
+        }
+
+        return std::static_pointer_cast<CV::Item>(std::make_shared<CV::List>(CV::List(result, false)));
+
+    }, false)));
+
+
+    ctx->set("l-cut", std::make_shared<CV::Function>(Function({"l", "p", "n"}, [](const std::vector<std::shared_ptr<CV::Item>> &params, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+        CV::FunctionConstraints consts;
+        consts.setMinParams(2);
+        consts.setMaxParams(3);
+        consts.allowMisMatchingTypes = true;
+        consts.allowNil = false;
+        consts.setExpectedTypeAt(0, CV::ItemTypes::LIST);
+        consts.setExpectedTypeAt(1, CV::ItemTypes::NUMBER);
+        consts.setExpectedTypeAt(2, CV::ItemTypes::NUMBER);
+        static const std::string name = "l-cut";
+        std::string errormsg;
+        if(!consts.test(params, errormsg)){
+            cursor->setError("'"+name+"': "+errormsg);
+            return std::make_shared<CV::Item>(CV::Item());
+        }
+
+        int index = std::static_pointer_cast<CV::Number>(params[1])->n;
+        auto list = std::static_pointer_cast<CV::List>(params[0]);
+        int n = params.size() > 2 ? std::static_pointer_cast<CV::Number>(params[2])->n : 1;
+
+        if(index+n < 0 || index+n > list->data.size()){
+            cursor->setError("'"+name+"': position("+std::to_string(index+n)+") is out of range("+std::to_string(list->data.size())+")");
+            return std::make_shared<CV::Item>(CV::Item());
+        }      
+
+        std::vector<std::shared_ptr<CV::Item>> result;
+        for(int i = 0; i < list->data.size(); ++i){
+            if(i >= index && i < index+n){
+                continue;
+            }
+            result.push_back(list->get(i));
+        }
+
+        return std::static_pointer_cast<CV::Item>(std::make_shared<CV::List>(CV::List(result, false)));
+
+
+    }, false)));      
+
+    ctx->set("l-reverse", std::make_shared<CV::Function>(CV::Function({}, [](const std::vector<std::shared_ptr<CV::Item>> &params, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+        CV::FunctionConstraints consts;
+        consts.setMinParams(1);
+        consts.setExpectType(CV::ItemTypes::LIST);
+        consts.allowMisMatchingTypes = false;
+        consts.allowNil = false;
+
+        std::string errormsg;
+        if(!consts.test(params, errormsg)){
+            cursor->setError("'l-reverse': "+errormsg);
+            return std::make_shared<CV::Item>(CV::Item());
+        }
+
+        auto reverse = [&](std::shared_ptr<CV::List> &lst){
+            std::vector<std::shared_ptr<CV::Item>> result;
+            for(int i = 0; i < lst->data.size(); ++i){
+                result.push_back(lst->get(lst->data.size()-1 - i));
+            }
+            return std::make_shared<CV::List>(CV::List(result, false));
+        };
+
+        if(params.size() == 1){
+            auto item = std::static_pointer_cast<CV::List>(params[0]);
+            return std::static_pointer_cast<CV::Item>(reverse(item));
+        }else{
+            std::vector<std::shared_ptr<CV::Item>> items;
+            for(int i = 0; i < params.size(); ++i){
+                auto item = std::static_pointer_cast<CV::List>(params[i]);
+                items.push_back(
+                    std::static_pointer_cast<CV::Item>(reverse(item))
+                );
+            }
+            return std::static_pointer_cast<CV::Item>(std::make_shared<CV::List>(CV::List(items, false)));
+        }
+
+
+
+    }, true)));        
+
+
+    /*
+        STRING
+    */
+    ctx->set("s-reverse", std::make_shared<CV::Function>(CV::Function({}, [](const std::vector<std::shared_ptr<CV::Item>> &params, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+        CV::FunctionConstraints consts;
+        consts.setMinParams(1);
+        consts.setExpectType(CV::ItemTypes::LIST);
+        consts.allowMisMatchingTypes = false;
+        consts.allowNil = false;
+
+        std::string errormsg;
+        if(!consts.test(params, errormsg)){
+            cursor->setError("'l-reverse': "+errormsg);
+            return std::make_shared<CV::Item>(CV::Item());
+        }
+
+        auto reverse = [&](std::shared_ptr<CV::String> &str){
+            std::string result;
+            for(int i = 0; i < str->data.size(); ++i){
+                result += str->data[str->data.size()-1 - i];
+            }
+            return result;
+        };
+
+        if(params.size() == 1){
+            auto casted = std::static_pointer_cast<CV::String>(params[0]);
+            return std::static_pointer_cast<CV::Item>(std::make_shared<CV::String>(CV::String(reverse(casted))));
+        }else{
+            std::vector<std::shared_ptr<CV::Item>> items;
+            for(int i = 0; i < params.size(); ++i){
+                auto casted = std::static_pointer_cast<CV::String>(params[i]);
+                items.push_back(
+                    std::static_pointer_cast<CV::Item>(std::make_shared<CV::String>(CV::String(reverse(casted))))
+                );
+            }
+            return std::static_pointer_cast<CV::Item>(std::make_shared<CV::List>(CV::List(items, false)));
+        }
+
+
+
+    }, true)));  
+
+    ctx->set("s-join", std::make_shared<CV::Function>(CV::Function({}, [](const std::vector<std::shared_ptr<CV::Item>> &params, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+        CV::FunctionConstraints consts;
+        // consts.setMinParams(2);
+        consts.allowMisMatchingTypes = false;
+        consts.allowNil = false;
+        consts.setExpectType(CV::ItemTypes::STRING);
+
+        static const std::string name = "s-join";
+        std::string errormsg;
+        if(!consts.test(params, errormsg)){
+            cursor->setError("'"+name+"': "+errormsg);
+            return std::make_shared<CV::Item>(CV::Item());
+        }
+        std::string result;
+        for(int i = 0; i < params.size(); ++i){
+            auto str = std::static_pointer_cast<CV::String>(params[i]);
+            result += str->get();
+        }
+        return std::static_pointer_cast<CV::Item>(std::make_shared<CV::String>(CV::String(result)));
+    }, true)));   
+
+    ctx->set("s-cut", std::make_shared<CV::Function>(Function({"s", "p", "n"}, [](const std::vector<std::shared_ptr<CV::Item>> &params, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+        CV::FunctionConstraints consts;
+        consts.setMinParams(2);
+        consts.setMaxParams(3);
+        consts.allowMisMatchingTypes = true;
+        consts.allowNil = false;
+        consts.setExpectedTypeAt(0, CV::ItemTypes::STRING);
+        consts.setExpectedTypeAt(1, CV::ItemTypes::NUMBER);
+        consts.setExpectedTypeAt(2, CV::ItemTypes::NUMBER);
+
+        static const std::string name = "s-cut";
+        std::string errormsg;
+        if(!consts.test(params, errormsg)){
+            cursor->setError("'"+name+"': "+errormsg);
+            return std::make_shared<CV::Item>(CV::Item());
+        }
+
+
+
+        int index = std::static_pointer_cast<CV::Number>(params[1])->n;
+        auto str = std::static_pointer_cast<CV::String>(params[0]);
+        int n = params.size() > 2 ? std::static_pointer_cast<CV::Number>(params[2])->n : 1;
+
+        if(index+n < 0 || index+n > str->data.size()){
+            cursor->setError("'"+name+"': position("+std::to_string(index+n)+") is out of range("+std::to_string(str->data.size())+")");
+            return std::make_shared<CV::Item>(CV::Item());
+        }      
+
+        std::string result;
+        for(int i = 0; i < str->data.size(); ++i){
+            if(i >= index && i < index+n){
+                continue;
+            }
+            result += str->data[i];
+        }
+
+        return std::static_pointer_cast<CV::Item>(std::make_shared<CV::String>(CV::String(result)));
+
+    }, false)));       
+
+
+    ctx->set("s-sub", std::make_shared<CV::Function>(Function({"s", "p", "n"}, [](const std::vector<std::shared_ptr<CV::Item>> &params, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+        CV::FunctionConstraints consts;
+        consts.setMinParams(2);
+        consts.setMaxParams(3);
+        consts.allowMisMatchingTypes = true;
+        consts.allowNil = false;
+        consts.setExpectedTypeAt(0, CV::ItemTypes::STRING);
+        consts.setExpectedTypeAt(1, CV::ItemTypes::NUMBER);
+        consts.setExpectedTypeAt(2, CV::ItemTypes::NUMBER);
+
+        static const std::string name = "s-sub";
+        std::string errormsg;
+        if(!consts.test(params, errormsg)){
+            cursor->setError("'"+name+"': "+errormsg);
+            return std::make_shared<CV::Item>(CV::Item());
+        }
+
+
+        int index = std::static_pointer_cast<CV::Number>(params[1])->n;
+        auto str = std::static_pointer_cast<CV::String>(params[0]);
+        int n = params.size() > 2 ? std::static_pointer_cast<CV::Number>(params[2])->n : 1;
+
+        if(index+n < 0 || index+n > str->data.size()){
+            cursor->setError("'"+name+"': position("+std::to_string(index+n)+") is out of range("+std::to_string(str->data.size())+")");
+            return std::make_shared<CV::Item>(CV::Item());
+        }      
+
+        std::string result;
+        for(int i = 0; i < str->data.size(); ++i){
+            if(i >= index && i < index+n){
+                result += str->data[i];
+            }
+        }
+
+        return std::static_pointer_cast<CV::Item>(std::make_shared<CV::String>(CV::String(result)));
+
+    }, false)));     
+
+
+    /*
+        LOGIC
+    */
+    ctx->set("or", std::make_shared<CV::Function>(CV::Function({}, [](const std::vector<std::shared_ptr<CV::Item>> &params, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+        CV::FunctionConstraints consts;
+        consts.setMinParams(1);
+        std::string errormsg;
+        if(!consts.test(params, errormsg)){
+            cursor->setError("'or': "+errormsg);
+            return std::make_shared<CV::Item>(CV::Item());
+        }
+
+        for(int i = 0; i < params.size(); ++i){
+            if( (params[i]->type == CV::ItemTypes::NUMBER && std::static_pointer_cast<CV::Number>(params[i])->n != 0) ||
+                (params[i]->type != CV::ItemTypes::NIL && params[i]->type != CV::ItemTypes::NUMBER)){
+                return std::static_pointer_cast<CV::Item>(std::make_shared<CV::Number>(CV::Number(1)));
+            }
+        }
+
+        return std::static_pointer_cast<CV::Item>(std::make_shared<CV::Number>(CV::Number(0)));
+    }, true)));        
+
+    ctx->set("and", std::make_shared<CV::Function>(CV::Function({}, [](const std::vector<std::shared_ptr<CV::Item>> &params, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+        CV::FunctionConstraints consts;
+        consts.setMinParams(1);
+        std::string errormsg;
+        if(!consts.test(params, errormsg)){
+            cursor->setError("'and': "+errormsg);
+            return std::make_shared<CV::Item>(CV::Item());
+        }
+
+        for(int i = 0; i < params.size(); ++i){
+            if( (params[i]->type == CV::ItemTypes::NUMBER && std::static_pointer_cast<CV::Number>(params[i])->n == 0) || (params[i]->type == CV::ItemTypes::NIL) ){
+                return std::static_pointer_cast<CV::Item>(std::make_shared<CV::Number>(CV::Number(0)));
+            }
+        }
+
+        return std::static_pointer_cast<CV::Item>(std::make_shared<CV::Number>(CV::Number(1)));
+    }, true)));    
+
+    ctx->set("not", std::make_shared<CV::Function>(CV::Function({}, [](const std::vector<std::shared_ptr<CV::Item>> &params, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+        CV::FunctionConstraints consts;
+        consts.setMinParams(1);
+        std::string errormsg;
+        if(!consts.test(params, errormsg)){
+            cursor->setError("'not': "+errormsg);
+            return std::make_shared<CV::Item>(CV::Item());
+        }
+        auto val = [](std::shared_ptr<CV::Item> &item){
+            if(item->type == CV::ItemTypes::NUMBER){
+                return std::static_pointer_cast<CV::Number>(item)->n == 0 ? std::make_shared<CV::Number>(CV::Number(1)) : std::make_shared<CV::Number>(CV::Number(0));
+            }else
+            if(item->type == CV::ItemTypes::NIL){
+                return std::make_shared<CV::Number>(CV::Number(1));
+            }else{
+                return std::make_shared<CV::Number>(CV::Number(0));
+            }
+        };
+        if(params.size() == 1){
+            auto item = params[0];
+            return std::static_pointer_cast<CV::Item>(val(item));
+        }else{
+            std::vector<std::shared_ptr<CV::Item>> results;
+            for(int i = 0; i < params.size(); ++i){
+                auto item = params[i];
+                results.push_back(std::static_pointer_cast<CV::Item>(val(item)));
+            }
+            return std::static_pointer_cast<CV::Item>(std::make_shared<CV::List>(CV::List(results)));
+        }
+        return std::make_shared<CV::Item>(CV::Item());
+    }, true)));                  
+
+
+    ctx->set("eq", std::make_shared<CV::Function>(CV::Function({}, [](const std::vector<std::shared_ptr<CV::Item>> &params, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+        CV::FunctionConstraints consts;
+        consts.setMinParams(2);
+        consts.allowMisMatchingTypes = false;
+        consts.allowNil = false;
+        std::string errormsg;
+        if(!consts.test(params, errormsg)){
+            cursor->setError("'eq': "+errormsg);
+            return std::make_shared<CV::Item>(CV::Item());
+        }
+        auto last = params[0];
+        for(int i = 1; i < params.size(); ++i){
+            auto item = params[i];
+            if(!last->isEq(item)){
+                return std::static_pointer_cast<CV::Item>(std::make_shared<CV::Number>(CV::Number(0)));
+            }
+        }				
+        return std::static_pointer_cast<CV::Item>(std::make_shared<CV::Number>(CV::Number(1)));
+    }, true)));     
+
+
+    ctx->set("neq", std::make_shared<CV::Function>(CV::Function({}, [](const std::vector<std::shared_ptr<CV::Item>> &params, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+        CV::FunctionConstraints consts;
+        consts.setMinParams(2);
+        consts.allowMisMatchingTypes = false;
+        consts.allowNil = false;
+        std::string errormsg;
+        if(!consts.test(params, errormsg)){
+            cursor->setError("'neq': "+errormsg);
+            return std::make_shared<CV::Item>(CV::Item());
+        }
+        for(int i = 0; i < params.size(); ++i){
+            auto p1 = params[i];
+            for(int j = 0; j < params.size(); ++j){
+                if(i == j){
+                    continue;
+                }
+                auto p2 = params[j];
+                if(p1->isEq(p2)){
+                    return std::static_pointer_cast<CV::Item>(std::make_shared<CV::Number>(CV::Number(0)));
+                }
+            }
+        }				
+        return std::static_pointer_cast<CV::Item>(std::make_shared<CV::Number>(CV::Number(1)));
+    }, true)));      
+
+
+    ctx->set("gt", std::make_shared<CV::Function>(CV::Function({}, [](const std::vector<std::shared_ptr<CV::Item>> &params, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+        CV::FunctionConstraints consts;
+        consts.setMinParams(2);
+        consts.allowNil = false;
+        consts.allowMisMatchingTypes = false;
+        consts.setExpectType(CV::ItemTypes::NUMBER);
+        std::string errormsg;
+        if(!consts.test(params, errormsg)){
+            cursor->setError("'gt': "+errormsg);
+            return std::make_shared<CV::Item>(CV::Item());
+        }
+        auto last = std::static_pointer_cast<CV::Number>(params[0])->n;
+        for(int i = 1; i < params.size(); ++i){
+            if(last <= std::static_pointer_cast<CV::Number>(params[i])->n){
+                return std::static_pointer_cast<CV::Item>(std::make_shared<CV::Number>(CV::Number(0)));
+            }
+            last = std::static_pointer_cast<CV::Number>(params[i])->n;
+        }				
+        return std::static_pointer_cast<CV::Item>(std::make_shared<CV::Number>(CV::Number(1)));
+    }, true)));    
+
+
+    ctx->set("gte", std::make_shared<CV::Function>(CV::Function({}, [](const std::vector<std::shared_ptr<CV::Item>> &params, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+        CV::FunctionConstraints consts;
+        consts.setMinParams(2);
+        consts.allowNil = false;
+        consts.allowMisMatchingTypes = false;
+        consts.setExpectType(CV::ItemTypes::NUMBER);
+        std::string errormsg;
+        if(!consts.test(params, errormsg)){
+            cursor->setError("'gte': "+errormsg);
+            return std::make_shared<CV::Item>(CV::Item());
+        }
+        auto last = std::static_pointer_cast<CV::Number>(params[0])->n;
+        for(int i = 1; i < params.size(); ++i){
+            if(last <  std::static_pointer_cast<CV::Number>(params[i])->n){
+                return std::static_pointer_cast<CV::Item>(std::make_shared<CV::Number>(CV::Number(0)));
+            }
+            last = std::static_pointer_cast<CV::Number>(params[i])->n;
+        }				
+        return std::static_pointer_cast<CV::Item>(std::make_shared<CV::Number>(CV::Number(1)));
+    }, true)));         
+
+
+    ctx->set("lt", std::make_shared<CV::Function>(CV::Function({}, [](const std::vector<std::shared_ptr<CV::Item>> &params, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+        CV::FunctionConstraints consts;
+        consts.setMinParams(2);
+        consts.allowNil = false;
+        consts.allowMisMatchingTypes = false;
+        consts.setExpectType(CV::ItemTypes::NUMBER);
+        std::string errormsg;
+        if(!consts.test(params, errormsg)){
+            cursor->setError("'lt': "+errormsg);
+            return std::make_shared<CV::Item>(CV::Item());
+        }
+        auto last = std::static_pointer_cast<CV::Number>(params[0])->n;
+        for(int i = 1; i < params.size(); ++i){
+            if(last >=  std::static_pointer_cast<CV::Number>(params[i])->n){
+                return std::static_pointer_cast<CV::Item>(std::make_shared<CV::Number>(CV::Number(0)));
+            }
+            last = std::static_pointer_cast<CV::Number>(params[i])->n;
+        }				
+        return std::static_pointer_cast<CV::Item>(std::make_shared<CV::Number>(CV::Number(1)));
+    }, true)));    
+
+    ctx->set("lte", std::make_shared<CV::Function>(CV::Function({}, [](const std::vector<std::shared_ptr<CV::Item>> &params, CV::Cursor *cursor, std::shared_ptr<CV::Context> &ctx){
+        CV::FunctionConstraints consts;
+        consts.setMinParams(2);
+        consts.allowNil = false;
+        consts.allowMisMatchingTypes = false;
+        consts.setExpectType(CV::ItemTypes::NUMBER);
+        std::string errormsg;
+        if(!consts.test(params, errormsg)){
+            cursor->setError("'gte': "+errormsg);
+            return std::make_shared<CV::Item>(CV::Item());
+        }
+        auto last = std::static_pointer_cast<CV::Number>(params[0])->n;
+        for(int i = 1; i < params.size(); ++i){
+            if(last >  std::static_pointer_cast<CV::Number>(params[i])->n){
+                return std::static_pointer_cast<CV::Item>(std::make_shared<CV::Number>(CV::Number(0)));
+            }
+            last = std::static_pointer_cast<CV::Number>(params[i])->n;
+        }				
+        return std::static_pointer_cast<CV::Item>(std::make_shared<CV::Number>(CV::Number(1)));
+    }, true)));    
+         
+}
